Split the spell handler's inline test module into scenario files (#624)
Fifteenth delivery of the refactor push, and the first outside session/.

handlers/spell.rs was the largest production file outside the Session tree at
6,283 lines, and its physical-policy entry had kept its split open since #584
C4 was written. Breaking it down by top-level item shows why: 3,909 of those
lines were a single inline mod tests. Sixty-two per cent of the file was
tests, and the production surface was only about 2,370 lines - so chasing the
production side first would have moved the smaller half.

The test module splits into bounded files under handlers/spell/tests,
following the handlers/trainer precedent: cast, mount, aura, loot_template,
totem, stored_item and misc for the 81 scenarios, plus fixtures, spell_stores
and port_fixtures for the 61 shared helpers. Every file is under the
1,000-line review budget. handlers/spell.rs drops from 6,283 to 2,375 lines
and its ceiling is tightened to that figure.

The test count is unchanged at 3,822 and not one production line moved. Test
code needed visibility widening the Session deliveries did not: shared
fixtures moved into a child module, so their definitions, struct fields and
inherent methods became pub(super) - trait-impl methods excepted, since Rust
does not permit a qualifier there.

Recorded, not fixed here: the remaining production half still holds methods
that are not spell operations - generate_item_loot_template_entries_like_cpp
at 205 lines, handle_open_item at 175, plus the loot-template row loaders and
the stored-item and wrapped-gift persistence. They belong to the loot and item
responsibilities, and moving them needs its own consumer migration.

Verified: cargo check --workspace --all-targets --locked clean, cargo test
-p wow-world --lib 3,822 passed and 0 failed, session-ownership-check
--syntax-only PASS, check_architecture.py check and self-test PASS, fmt and
diff hygiene clean.
