+++ crates/wow-world/src/handlers/spell.rs
@@ -2371,3912 +2371,5 @@ fn add_loot_template_row_item_like_cpp<F>(
 }
 
 #[cfg(test)]
-mod tests {
-    use std::collections::VecDeque;
-    use std::sync::{Arc, Mutex};
-    use std::time::Instant;
-
-    use rand::{Rng, SeedableRng, rngs::StdRng};
-
-    use wow_constants::{
-        BagFamilyMask, DeathState, ItemContext, ItemFieldFlags, ItemFlags, ItemUpdateState,
-        PowerType, ServerOpcodes, SpellCastResult,
-    };
-    use wow_core::{ObjectGuid, Position, guid::HighGuid};
-    use wow_entities::{
-        AppliedAuraRef, Creature, Item, ItemCreateInfo, MAX_ITEM_SPELLS, Pet, PetType, Player,
-        UNIT_MASK_TOTEM,
-    };
-    use wow_loot::{
-        LootConditionRowLikeCpp, condition_compare_values_like_cpp,
-        loot_conditions_allow_player_like_cpp_representable,
-    };
-    use wow_packet::WorldPacket;
-    use wow_packet::packets::loot::{LootEntry, LootEntryFlags};
-    use wow_packet::packets::movement::MovementInfo;
-    use wow_packet::packets::spell::{SpellCastVisual, SpellTargetData};
-    use wow_persistence::{
-        ItemTemplateAddonCatalogPersistencePortLikeCpp, ItemTemplateAddonCatalogRequestLikeCpp,
-        ItemTemplateAddonLootMetadataOutcomeLikeCpp, ItemTemplateAddonMoneyOutcomeLikeCpp,
-        PersistenceFutureLikeCpp,
-    };
-
-    use super::{
-        ITEM_FLAGS_CU_FOLLOW_LOOT_RULES_LIKE_CPP, ITEM_FLAGS_CU_IGNORE_QUEST_STATUS_LIKE_CPP,
-        ItemTemplateAddonLootMetadataLikeCpp, LOOT_MODE_DEFAULT_LIKE_CPP, LootTemplateRow,
-        add_loot_item_stacks_like_cpp, add_loot_template_row_item_like_cpp,
-        apply_wrapped_gift_transform_like_cpp, item_loot_quest_status_allows_like_cpp,
-        loot_entry_flags_for_row_metadata_like_cpp, loot_template_group_row_can_roll_like_cpp,
-        loot_template_plain_row_can_roll_like_cpp, loot_template_reference_row_can_roll_like_cpp,
-        normalize_item_money_loot_bounds_like_cpp, player_class_mask_like_cpp,
-        player_quest_status_mask_like_cpp, player_race_mask_like_cpp,
-        referenced_loot_max_count_like_cpp, roll_chance_with_rate_like_cpp,
-        roll_group_loot_row_like_cpp, stored_item_row_can_load_like_cpp_representable,
-        stored_loot_item_should_persist_like_cpp,
-    };
-    use crate::session::{
-        AuraApplication, RepresentedAuraEffectLikeCpp, RepresentedPendingSpellCastRequestLikeCpp,
-        SessionPlayerController, SharedCanonicalMapManager, SpellCastMetadata, SpellCastState,
-    };
-
-    fn make_session() -> (crate::session::WorldSession, flume::Receiver<Vec<u8>>) {
-        let (_pkt_tx, pkt_rx) = flume::bounded(100);
-        let (send_tx, send_rx) = flume::bounded(100);
-
-        (
-            crate::session::WorldSession::new(
-                1,
-                "TestAccount".into(),
-                0,
-                2,
-                9,
-                54261,
-                vec![0u8; 40],
-                "esES".into(),
-                pkt_rx,
-                send_tx,
-            ),
-            send_rx,
-        )
-    }
-
-    struct ItemTemplateAddonCatalogPortFixtureLikeCpp {
-        money_requests: Mutex<Vec<ItemTemplateAddonCatalogRequestLikeCpp>>,
-        money_outcomes: Mutex<VecDeque<ItemTemplateAddonMoneyOutcomeLikeCpp>>,
-        metadata_requests: Mutex<Vec<ItemTemplateAddonCatalogRequestLikeCpp>>,
-        metadata_outcomes: Mutex<VecDeque<ItemTemplateAddonLootMetadataOutcomeLikeCpp>>,
-    }
-
-    impl ItemTemplateAddonCatalogPortFixtureLikeCpp {
-        fn new(
-            money_outcomes: impl IntoIterator<Item = ItemTemplateAddonMoneyOutcomeLikeCpp>,
-            metadata_outcomes: impl IntoIterator<Item = ItemTemplateAddonLootMetadataOutcomeLikeCpp>,
-        ) -> Arc<Self> {
-            Arc::new(Self {
-                money_requests: Mutex::new(Vec::new()),
-                money_outcomes: Mutex::new(money_outcomes.into_iter().collect()),
-                metadata_requests: Mutex::new(Vec::new()),
-                metadata_outcomes: Mutex::new(metadata_outcomes.into_iter().collect()),
-            })
-        }
-    }
-
-    impl ItemTemplateAddonCatalogPersistencePortLikeCpp for ItemTemplateAddonCatalogPortFixtureLikeCpp {
-        fn load_item_template_addon_money_like_cpp<'a>(
-            &'a self,
-            request: ItemTemplateAddonCatalogRequestLikeCpp,
-        ) -> PersistenceFutureLikeCpp<'a, ItemTemplateAddonMoneyOutcomeLikeCpp> {
-            self.money_requests.lock().unwrap().push(request);
-            let outcome = self
-                .money_outcomes
-                .lock()
-                .unwrap()
-                .pop_front()
-                .expect("one item-addon money outcome per request");
-            Box::pin(async move { outcome })
-        }
-
-        fn load_item_template_addon_loot_metadata_like_cpp<'a>(
-            &'a self,
-            request: ItemTemplateAddonCatalogRequestLikeCpp,
-        ) -> PersistenceFutureLikeCpp<'a, ItemTemplateAddonLootMetadataOutcomeLikeCpp> {
-            self.metadata_requests.lock().unwrap().push(request);
-            let outcome = self
-                .metadata_outcomes
-                .lock()
-                .unwrap()
-                .pop_front()
-                .expect("one item-addon metadata outcome per request");
-            Box::pin(async move { outcome })
-        }
-    }
-
-    fn shared_canonical_map_manager() -> SharedCanonicalMapManager {
-        Arc::new(Mutex::new(wow_map::MapManager::default()))
-    }
-
-    fn add_canonical_test_player_on_map(
-        canonical: &SharedCanonicalMapManager,
-        guid: ObjectGuid,
-        position: Position,
-        map_id: u32,
-        instance_id: u32,
-    ) {
-        let mut player = Player::new(Some(1), false);
-        player.unit_mut().world_mut().object_mut().create(guid);
-        player.unit_mut().world_mut().set_name("SpellHandlerPlayer");
-        player
-            .unit_mut()
-            .world_mut()
-            .set_map(map_id, instance_id)
-            .unwrap();
-        player.unit_mut().world_mut().relocate(position);
-        player.unit_mut().world_mut().object_mut().add_to_world();
-
-        canonical
-            .lock()
-            .unwrap()
-            .create_world_map(map_id, instance_id)
-            .map_mut()
-            .insert_map_object_record(wow_entities::MapObjectRecord::new_player(player).unwrap())
-            .unwrap();
-    }
-
-    fn install_canonical_player(
-        session: &mut crate::session::WorldSession,
-        canonical: &SharedCanonicalMapManager,
-        player_guid: ObjectGuid,
-    ) {
-        let position = Position::new(10.0, 20.0, 30.0, 0.0);
-        session.set_canonical_map_manager(Arc::clone(canonical));
-        session.attach_player_controller_like_cpp(SessionPlayerController::new(
-            player_guid,
-            "SpellHandlerPlayer".to_string(),
-            position,
-            571,
-            1,
-            1,
-            80,
-            0,
-        ));
-        add_canonical_test_player_on_map(canonical, player_guid, position, 571, 0);
-        assert!(session.adopt_registered_canonical_player_fixture_like_cpp());
-        session.set_player_moved_unit_guid_like_cpp(player_guid);
-        session.set_legacy_creature_aggro_config_like_cpp(
-            crate::session::LegacyCreatureAggroConfigLikeCpp {
-                spell_x_spell_visual_store: Some(Arc::new(
-                    wow_data::SpellXSpellVisualStore::from_entries([]),
-                )),
-                ..Default::default()
-            },
-        );
-    }
-
-    fn add_canonical_test_pet_on_map(
-        canonical: &SharedCanonicalMapManager,
-        owner_guid: ObjectGuid,
-        pet_guid: ObjectGuid,
-        spell_id: u32,
-        alive: bool,
-    ) {
-        let mut pet = Pet::new(owner_guid, PetType::Hunter);
-        pet.creature_mut()
-            .unit_mut()
-            .world_mut()
-            .object_mut()
-            .create(pet_guid);
-        pet.creature_mut()
-            .unit_mut()
-            .world_mut()
-            .set_map(571, 0)
-            .unwrap();
-        pet.creature_mut()
-            .unit_mut()
-            .world_mut()
-            .relocate(Position::new(10.5, 20.5, 30.0, 0.0));
-        pet.creature_mut()
-            .unit_mut()
-            .world_mut()
-            .object_mut()
-            .add_to_world();
-        pet.creature_mut().unit_mut().set_max_health(100);
-        pet.creature_mut()
-            .unit_mut()
-            .set_health(if alive { 100 } else { 0 });
-        if !alive {
-            pet.creature_mut()
-                .unit_mut()
-                .set_death_state(DeathState::Dead);
-        }
-        let aura = AppliedAuraRef::new(spell_id, ObjectGuid::EMPTY, 0, 0x1);
-        pet.creature_mut()
-            .unit_mut()
-            .subsystems_mut()
-            .auras
-            .add_applied(aura);
-
-        canonical
-            .lock()
-            .unwrap()
-            .find_map_mut(571, 0)
-            .unwrap()
-            .map_mut()
-            .insert_map_object_record(wow_entities::MapObjectRecord::new_pet(pet).unwrap())
-            .unwrap();
-    }
-
-    fn add_canonical_test_creature_on_map(
-        canonical: &SharedCanonicalMapManager,
-        guid: ObjectGuid,
-        position: Position,
-        map_id: u32,
-        instance_id: u32,
-        is_totem: bool,
-    ) {
-        let mut creature = Creature::new(false);
-        creature.unit_mut().world_mut().object_mut().create(guid);
-        creature.unit_mut().world_mut().object_mut().set_entry(777);
-        creature
-            .unit_mut()
-            .world_mut()
-            .set_map(map_id, instance_id)
-            .unwrap();
-        creature.unit_mut().world_mut().relocate(position);
-        creature.unit_mut().world_mut().object_mut().add_to_world();
-        if is_totem {
-            creature.add_unit_type_mask_like_cpp(UNIT_MASK_TOTEM);
-        }
-
-        canonical
-            .lock()
-            .unwrap()
-            .find_map_mut(map_id, instance_id)
-            .unwrap()
-            .map_mut()
-            .insert_map_object_record(
-                wow_entities::MapObjectRecord::new_creature(creature).unwrap(),
-            )
-            .unwrap();
-    }
-
-    fn set_canonical_player_pet_guid(
-        canonical: &SharedCanonicalMapManager,
-        player_guid: ObjectGuid,
-        pet_guid: ObjectGuid,
-    ) {
-        canonical
-            .lock()
-            .unwrap()
-            .find_map_mut(571, 0)
-            .unwrap()
-            .map_mut()
-            .get_typed_player_mut(player_guid)
-            .unwrap()
-            .unit_mut()
-            .subsystems_mut()
-            .control
-            .set_pet_guid(pet_guid);
-    }
-
-    fn set_canonical_player_charmed_guid(
-        canonical: &SharedCanonicalMapManager,
-        player_guid: ObjectGuid,
-        charmed_guid: ObjectGuid,
-    ) {
-        canonical
-            .lock()
-            .unwrap()
-            .find_map_mut(571, 0)
-            .unwrap()
-            .map_mut()
-            .get_typed_player_mut(player_guid)
-            .unwrap()
-            .unit_mut()
-            .subsystems_mut()
-            .control
-            .charmed_guid = Some(charmed_guid);
-    }
-
-    fn canonical_pet_has_applied_aura(
-        canonical: &SharedCanonicalMapManager,
-        pet_guid: ObjectGuid,
-        aura: AppliedAuraRef,
-    ) -> bool {
-        canonical
-            .lock()
-            .unwrap()
-            .find_map(571, 0)
-            .unwrap()
-            .map()
-            .get_typed_pet(pet_guid)
-            .unwrap()
-            .creature()
-            .unit()
-            .subsystems()
-            .auras
-            .has_applied(aura)
-    }
-
-    fn canonical_creature_has_applied_aura(
-        canonical: &SharedCanonicalMapManager,
-        creature_guid: ObjectGuid,
-        aura: AppliedAuraRef,
-    ) -> bool {
-        canonical
-            .lock()
-            .unwrap()
-            .find_map(571, 0)
-            .unwrap()
-            .map()
-            .with_creature_like_cpp(creature_guid, |creature| {
-                creature.unit().subsystems().auras.has_applied(aura)
-            })
-            .unwrap()
-    }
-
-    fn set_canonical_player_summon_slot(
-        canonical: &SharedCanonicalMapManager,
-        player_guid: ObjectGuid,
-        slot: usize,
-        guid: ObjectGuid,
-    ) {
-        canonical
-            .lock()
-            .unwrap()
-            .find_map_mut(571, 0)
-            .unwrap()
-            .map_mut()
-            .get_typed_player_mut(player_guid)
-            .unwrap()
-            .unit_mut()
-            .subsystems_mut()
-            .control
-            .set_summon_slot(slot, guid);
-    }
-
-    fn canonical_player_summon_slot(
-        canonical: &SharedCanonicalMapManager,
-        player_guid: ObjectGuid,
-        slot: usize,
-    ) -> ObjectGuid {
-        canonical
-            .lock()
-            .unwrap()
-            .find_map(571, 0)
-            .unwrap()
-            .map()
-            .get_typed_player(player_guid)
-            .unwrap()
-            .unit()
-            .subsystems()
-            .control
-            .summon_slots[slot]
-    }
-
-    fn canonical_creature_exists(canonical: &SharedCanonicalMapManager, guid: ObjectGuid) -> bool {
-        canonical
-            .lock()
-            .unwrap()
-            .find_map(571, 0)
-            .unwrap()
-            .map()
-            .with_creature_like_cpp(guid, |_| ())
-            .is_some()
-    }
-
-    fn install_active_spell_cast(
-        session: &mut crate::session::WorldSession,
-        spell_id: i32,
-        cast_id: ObjectGuid,
-    ) {
-        let player_guid = ObjectGuid::create_player(1, 42);
-        session.set_active_spell_cast_like_cpp(Some(SpellCastState {
-            spell_id,
-            target_guid: player_guid,
-            target_data: wow_entities::SpellCastTargetsLikeCpp {
-                flags: 0x2, // SpellCastTargetFlags::Unit
-                unit: player_guid,
-                ..Default::default()
-            },
-            cast_id,
-            cast_start_time: std::time::Instant::now(),
-            cast_time_ms: 30_000,
-            spell_visual: wow_entities::SpellCastVisualLikeCpp {
-                spell_visual_id: 1,
-                script_visual_id: 0,
-            },
-            metadata: SpellCastMetadata::default(),
-        }));
-    }
-
-    fn install_pending_spell_cast_request(
-        session: &mut crate::session::WorldSession,
-        spell_id: i32,
-        cast_id: ObjectGuid,
-    ) {
-        session.request_represented_spell_cast_like_cpp(
-            RepresentedPendingSpellCastRequestLikeCpp {
-                cast_id,
-                spell_id,
-                casting_unit_guid: ObjectGuid::create_player(1, 42),
-                target_guid: ObjectGuid::create_player(1, 42),
-                target_data: wow_entities::SpellCastTargetsLikeCpp {
-                    flags: 0x2,
-                    unit: ObjectGuid::create_player(1, 42),
-                    ..Default::default()
-                },
-                spell_visual: wow_entities::SpellCastVisualLikeCpp {
-                    spell_visual_id: 0,
-                    script_visual_id: 0,
-                },
-                metadata: SpellCastMetadata::default(),
-            },
-        );
-    }
-
-    fn install_canonical_channeled_spell(
-        session: &mut crate::session::WorldSession,
-        player_guid: ObjectGuid,
-        spell_id: u32,
-    ) -> wow_entities::CurrentSpellRef {
-        let spell = wow_entities::CurrentSpellRef::new(spell_id, Some(player_guid), None)
-            .with_state(wow_constants::SpellState::Delayed);
-        session.mutate_canonical_player_like_cpp(|player| {
-            player
-                .unit_mut()
-                .set_current_cast_spell(wow_entities::CurrentSpellSlot::Channeled, spell);
-        });
-        spell
-    }
-
-    fn canonical_channeled_spell_id(session: &mut crate::session::WorldSession) -> Option<u32> {
-        session
-            .mutate_canonical_player_like_cpp(|player| {
-                player
-                    .unit()
-                    .current_spell(wow_entities::CurrentSpellSlot::Channeled)
-                    .map(|spell| spell.spell_id)
-            })
-            .flatten()
-    }
-
-    fn cancel_cast_packet(cast_id: ObjectGuid, spell_id: u32) -> WorldPacket {
-        let mut pkt = WorldPacket::new_empty();
-        pkt.write_packed_guid(&cast_id);
-        pkt.write_uint32(spell_id);
-        pkt.reset_read();
-        pkt
-    }
-
-    fn cancel_channelling_packet(channel_spell: i32, reason: i32) -> WorldPacket {
-        let mut pkt = WorldPacket::new_empty();
-        pkt.write_int32(channel_spell);
-        pkt.write_int32(reason);
-        pkt.reset_read();
-        pkt
-    }
-
-    fn cast_spell_packet(spell_id: i32, caster_guid: ObjectGuid) -> WorldPacket {
-        cast_spell_packet_with_move_update(spell_id, caster_guid, None)
-    }
-
-    fn cast_spell_packet_with_move_update(
-        spell_id: i32,
-        caster_guid: ObjectGuid,
-        move_update: Option<MovementInfo>,
-    ) -> WorldPacket {
-        let mut pkt = WorldPacket::new_empty();
-        pkt.write_packed_guid(&caster_guid);
-        pkt.write_int32(0);
-        pkt.write_int32(0);
-        pkt.write_int32(spell_id);
-        SpellCastVisual {
-            spell_visual_id: 0,
-            script_visual_id: 0,
-        }
-        .write(&mut pkt);
-        pkt.write_float(0.0);
-        pkt.write_float(0.0);
-        pkt.write_packed_guid(&ObjectGuid::EMPTY);
-        pkt.write_uint32(0);
-        pkt.write_uint32(0);
-        pkt.write_uint32(0);
-        pkt.write_bits(0, 5);
-        pkt.write_bit(move_update.is_some());
-        pkt.write_bits(0, 2);
-        pkt.write_bit(false);
-        pkt.flush_bits();
-        SpellTargetData {
-            flags: 0x2,
-            unit: caster_guid,
-            ..SpellTargetData::default()
-        }
-        .write(&mut pkt);
-        if let Some(move_update) = move_update {
-            move_update.write(&mut pkt);
-        }
-        pkt.reset_read();
-        pkt
-    }
-
-    fn basic_spell_store(spell_ids: impl IntoIterator<Item = i32>) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        for spell_id in spell_ids {
-            spell_store.insert(
-                spell_id,
-                wow_data::SpellInfo {
-                    spell_id,
-                    cast_time_ms: 0,
-                    cooldown_ms: 0,
-                    recovery_time_ms: 0,
-                    effect_type: 0,
-                    effect_base_points: 0,
-                    effect_bonus_coefficient: 0.0,
-                    aura_type: None,
-                    display_flags: 0,
-                    requires_spell_focus: 0,
-                    power_costs: Vec::new(),
-                    effects: Vec::new(),
-                },
-            );
-        }
-        Arc::new(spell_store)
-    }
-
-    fn spell_store_with_mana_power_cost_like_cpp(
-        spell_id: i32,
-        mana_cost: i32,
-        power_cost_pct: f32,
-    ) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            spell_id,
-            wow_data::SpellInfo {
-                spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: 0,
-                effect_base_points: 0,
-                effect_bonus_coefficient: 0.0,
-                aura_type: None,
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: vec![wow_data::SpellPowerCostInfoLikeCpp {
-                    order_index: 0,
-                    power_type: PowerType::Mana as i8,
-                    mana_cost,
-                    mana_cost_per_level: 0,
-                    mana_per_second: 0,
-                    power_cost_pct,
-                    power_cost_max_pct: 0.0,
-                    power_pct_per_second: 0.0,
-                    required_aura_spell_id: 0,
-                    optional_cost: 0,
-                }],
-                effects: Vec::new(),
-            },
-        );
-        Arc::new(spell_store)
-    }
-
-    fn spell_store_with_global_cooldown_and_mana_power_cost_like_cpp(
-        spell_id: i32,
-        cooldown_ms: u32,
-        mana_cost: i32,
-        power_cost_pct: f32,
-    ) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            spell_id,
-            wow_data::SpellInfo {
-                spell_id,
-                cast_time_ms: 0,
-                cooldown_ms,
-                recovery_time_ms: 0,
-                effect_type: 0,
-                effect_base_points: 0,
-                effect_bonus_coefficient: 0.0,
-                aura_type: None,
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: vec![wow_data::SpellPowerCostInfoLikeCpp {
-                    order_index: 0,
-                    power_type: PowerType::Mana as i8,
-                    mana_cost,
-                    mana_cost_per_level: 0,
-                    mana_per_second: 0,
-                    power_cost_pct,
-                    power_cost_max_pct: 0.0,
-                    power_pct_per_second: 0.0,
-                    required_aura_spell_id: 0,
-                    optional_cost: 0,
-                }],
-                effects: Vec::new(),
-            },
-        );
-        Arc::new(spell_store)
-    }
-
-    fn spell_store_with_mana_cost_and_missing_focus_like_cpp(
-        spell_id: i32,
-        mana_cost: i32,
-    ) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            spell_id,
-            wow_data::SpellInfo {
-                spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: wow_data::spell::spell_effect_types::SPELL_EFFECT_SUMMON_OBJECT_WILD,
-                effect_base_points: 0,
-                effect_bonus_coefficient: 0.0,
-                aura_type: None,
-                display_flags: 0,
-                requires_spell_focus: 777,
-                power_costs: vec![wow_data::SpellPowerCostInfoLikeCpp {
-                    order_index: 0,
-                    power_type: PowerType::Mana as i8,
-                    mana_cost,
-                    mana_cost_per_level: 0,
-                    mana_per_second: 0,
-                    power_cost_pct: 0.0,
-                    power_cost_max_pct: 0.0,
-                    power_pct_per_second: 0.0,
-                    required_aura_spell_id: 0,
-                    optional_cost: 0,
-                }],
-                effects: vec![wow_data::SpellEffectInfo {
-                    effect_index: 0,
-                    effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_SUMMON_OBJECT_WILD,
-                    ..Default::default()
-                }],
-            },
-        );
-        Arc::new(spell_store)
-    }
-
-    fn set_canonical_player_mana_like_cpp(
-        session: &mut crate::session::WorldSession,
-        current: i32,
-        max: i32,
-    ) {
-        session.mutate_canonical_player_like_cpp(|player| {
-            player.set_power_index(PowerType::Mana, Some(0));
-            player.unit_mut().set_create_mana_like_cpp(max);
-            player.unit_mut().set_max_power(PowerType::Mana, max);
-            player.unit_mut().set_power(PowerType::Mana, current);
-        });
-    }
-
-    fn canonical_player_mana_like_cpp(session: &mut crate::session::WorldSession) -> i32 {
-        session
-            .mutate_canonical_player_like_cpp(|player| player.get_power(PowerType::Mana))
-            .unwrap_or(0)
-    }
-
-    fn canonical_player_power_like_cpp(
-        session: &mut crate::session::WorldSession,
-        power: PowerType,
-    ) -> i32 {
-        session
-            .mutate_canonical_player_like_cpp(|player| player.get_power(power))
-            .unwrap_or(0)
-    }
-
-    fn spell_store_with_global_cooldown(
-        spell_id: i32,
-        cooldown_ms: u32,
-    ) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            spell_id,
-            wow_data::SpellInfo {
-                spell_id,
-                cast_time_ms: 0,
-                cooldown_ms,
-                recovery_time_ms: 0,
-                effect_type: 0,
-                effect_base_points: 0,
-                effect_bonus_coefficient: 0.0,
-                aura_type: None,
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: Vec::new(),
-            },
-        );
-        Arc::new(spell_store)
-    }
-
-    fn mounted_spell_store(spell_id: i32, creature_entry: i32) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            spell_id,
-            wow_data::SpellInfo {
-                spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                effect_base_points: 77,
-                effect_bonus_coefficient: 0.0,
-                aura_type: Some(wow_data::spell::aura_types::SPELL_AURA_MOUNTED),
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: vec![wow_data::SpellEffectInfo {
-                    effect_index: 0,
-                    effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                    effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOUNTED,
-                    effect_base_points: 77,
-                    effect_misc_value_1: creature_entry,
-                    ..Default::default()
-                }],
-            },
-        );
-        Arc::new(spell_store)
-    }
-
-    fn mounted_flying_spell_store(spell_id: i32, creature_entry: i32) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            spell_id,
-            wow_data::SpellInfo {
-                spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                effect_base_points: 77,
-                effect_bonus_coefficient: 0.0,
-                aura_type: Some(wow_data::spell::aura_types::SPELL_AURA_MOUNTED),
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: vec![
-                    wow_data::SpellEffectInfo {
-                        effect_index: 0,
-                        effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                        effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOUNTED,
-                        effect_base_points: 77,
-                        effect_misc_value_1: creature_entry,
-                        ..Default::default()
-                    },
-                    wow_data::SpellEffectInfo {
-                        effect_index: 1,
-                        effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                        effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOD_INCREASE_MOUNTED_FLIGHT_SPEED,
-                        effect_base_points: 280,
-                        ..Default::default()
-                    },
-                ],
-            },
-        );
-        Arc::new(spell_store)
-    }
-
-    fn shapeshift_spell_store(spell_id: i32, form_id: i32) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            spell_id,
-            wow_data::SpellInfo {
-                spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                effect_base_points: 0,
-                effect_bonus_coefficient: 0.0,
-                aura_type: Some(wow_data::spell::aura_types::SPELL_AURA_MOD_SHAPESHIFT),
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: vec![wow_data::SpellEffectInfo {
-                    effect_index: 0,
-                    effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                    effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOD_SHAPESHIFT,
-                    effect_misc_value_1: form_id,
-                    ..Default::default()
-                }],
-            },
-        );
-        Arc::new(spell_store)
-    }
-
-    fn mounted_spell_store_with_active_shapeshift_aura(
-        mount_spell_id: i32,
-        shapeshift_spell_id: i32,
-        form_id: i32,
-    ) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            mount_spell_id,
-            wow_data::SpellInfo {
-                spell_id: mount_spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                effect_base_points: 77,
-                effect_bonus_coefficient: 0.0,
-                aura_type: Some(wow_data::spell::aura_types::SPELL_AURA_MOUNTED),
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: vec![wow_data::SpellEffectInfo {
-                    effect_index: 0,
-                    effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                    effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOUNTED,
-                    effect_base_points: 77,
-                    ..Default::default()
-                }],
-            },
-        );
-        spell_store.insert(
-            shapeshift_spell_id,
-            wow_data::SpellInfo {
-                spell_id: shapeshift_spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                effect_base_points: 0,
-                effect_bonus_coefficient: 0.0,
-                aura_type: Some(wow_data::spell::aura_types::SPELL_AURA_MOD_SHAPESHIFT),
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: vec![wow_data::SpellEffectInfo {
-                    effect_index: 0,
-                    effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                    effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOD_SHAPESHIFT,
-                    effect_misc_value_1: form_id,
-                    ..Default::default()
-                }],
-            },
-        );
-        Arc::new(spell_store)
-    }
-
-    fn mounted_spell_store_with_transform_spell(
-        mount_spell_id: i32,
-        transform_spell_id: i32,
-        allow_while_mounted: bool,
-    ) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            mount_spell_id,
-            wow_data::SpellInfo {
-                spell_id: mount_spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                effect_base_points: 77,
-                effect_bonus_coefficient: 0.0,
-                aura_type: Some(wow_data::spell::aura_types::SPELL_AURA_MOUNTED),
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: vec![wow_data::SpellEffectInfo {
-                    effect_index: 0,
-                    effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                    effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOUNTED,
-                    effect_base_points: 77,
-                    ..Default::default()
-                }],
-            },
-        );
-        spell_store.insert(
-            transform_spell_id,
-            wow_data::SpellInfo {
-                spell_id: transform_spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                effect_base_points: 0,
-                effect_bonus_coefficient: 0.0,
-                aura_type: Some(wow_data::spell::aura_types::SPELL_AURA_TRANSFORM),
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: vec![wow_data::SpellEffectInfo {
-                    effect_index: 0,
-                    effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                    effect_aura: wow_data::spell::aura_types::SPELL_AURA_TRANSFORM,
-                    ..Default::default()
-                }],
-            },
-        );
-        if allow_while_mounted {
-            let mut attributes = [0; 15];
-            attributes[0] = wow_data::spell::attributes::SPELL_ATTR0_ALLOW_WHILE_MOUNTED;
-            spell_store.insert_spell_misc_attributes_like_cpp(transform_spell_id, attributes);
-        }
-        Arc::new(spell_store)
-    }
-
-    fn active_shapeshift_aura_for_test(spell_id: i32, caster_guid: ObjectGuid) -> AuraApplication {
-        AuraApplication {
-            spell_id,
-            difficulty_id: 0,
-            caster_guid,
-            slot: 0,
-            duration_total: 0,
-            duration_remaining: 0,
-            stack_count: 1,
-            aura_flags: 0x0000_0001,
-            effect_mask: 1,
-            aura_interrupt_flags: 0,
-            aura_interrupt_flags2: 0,
-            represented_effect: None,
-            represented_amount: 0,
-            represented_effect_amounts: Vec::new(),
-            represented_misc_value: None,
-            represented_multiplier: 1.0,
-            applied_at: Instant::now(),
-        }
-    }
-
-    fn set_canonical_player_display_for_test(
-        canonical: &SharedCanonicalMapManager,
-        player_guid: ObjectGuid,
-        display_id: u32,
-        set_native: bool,
-    ) {
-        let mut manager = canonical.lock().unwrap();
-        let player = manager
-            .find_map_mut(571, 0)
-            .unwrap()
-            .map_mut()
-            .get_typed_player_mut(player_guid)
-            .unwrap();
-        player.unit_mut().set_display_id(display_id, set_native);
-    }
-
-    fn creature_display_info_extra_for_test(
-        id: u32,
-        display_race_id: i8,
-    ) -> wow_data::CreatureDisplayInfoExtraEntry {
-        wow_data::CreatureDisplayInfoExtraEntry {
-            id,
-            display_race_id,
-            display_sex_id: 0,
-            display_class_id: 0,
-            skin_id: 0,
-            face_id: 0,
-            hair_style_id: 0,
-            hair_color_id: 0,
-            facial_hair_id: 0,
-            flags: 0,
-            bake_material_resources_id: 0,
-            hd_bake_material_resources_id: 0,
-            custom_display_option: [0; 3],
-        }
-    }
-
-    fn chr_races_entry_for_test(
-        id: u32,
-        flags: i32,
-    ) -> wow_data::character_progression::ChrRacesEntry {
-        wow_data::character_progression::ChrRacesEntry {
-            id,
-            client_prefix: String::new(),
-            client_file_string: String::new(),
-            name: String::new(),
-            flags,
-            male_display_id: 0,
-            female_display_id: 0,
-            high_res_male_display_id: 0,
-            high_res_female_display_id: 0,
-            res_sickness_spell_id: 0,
-            splash_sound_id: 0,
-            create_screen_file_data_id: 0,
-            select_screen_file_data_id: 0,
-            low_res_screen_file_data_id: 0,
-            altered_form_start_visual_kit_id: [0; 3],
-            altered_form_finish_visual_kit_id: [0; 3],
-            heritage_armor_achievement_id: 0,
-            starting_level: 1,
-            ui_display_order: 0,
-            playable_race_bit: 0,
-            female_skeleton_file_data_id: 0,
-            male_skeleton_file_data_id: 0,
-            helmet_anim_scaling_race_id: 0,
-            transmogrify_disabled_slot_mask: 0,
-            faction_id: 0,
-            cinematic_sequence_id: 0,
-            base_language: 0,
-            creature_type: 0,
-            alliance: 0,
-            race_related: 0,
-            unaltered_visual_race_id: 0,
-            default_class_id: 0,
-            neutral_race_id: 0,
-        }
-    }
-
-    fn set_transformed_display_mount_check_stores_for_test(
-        session: &mut crate::session::WorldSession,
-        transformed_display_id: u32,
-        model_flags: u32,
-        race_flags: i32,
-    ) {
-        let display_extra_id = 91;
-        let model_id = 92;
-        let race_id = 7;
-        session.set_creature_display_info_store(Arc::new(
-            wow_data::CreatureDisplayInfoStore::from_entries([
-                wow_data::CreatureDisplayInfoEntry {
-                    id: transformed_display_id,
-                    model_id,
-                    extended_display_info_id: display_extra_id,
-                    creature_model_scale: 1.0,
-                },
-            ]),
-        ));
-        session.set_creature_display_info_extra_store(Arc::new(
-            wow_data::CreatureDisplayInfoExtraStore::from_entries([
-                creature_display_info_extra_for_test(display_extra_id as u32, race_id),
-            ]),
-        ));
-        session.set_creature_model_data_store(Arc::new(
-            wow_data::CreatureModelDataStore::from_entries([wow_data::CreatureModelDataEntry {
-                id: u32::from(model_id),
-                flags: model_flags,
-                file_data_id: 0,
-                collision_height: 2.0,
-                hover_height: 1.0,
-                model_scale: 1.0,
-                mount_height: 0.0,
-            }]),
-        ));
-        session.set_chr_races_store(Arc::new(
-            wow_data::character_progression::ChrRacesStore::from_entries([
-                chr_races_entry_for_test(u32::from(race_id as u8), race_flags),
-            ]),
-        ));
-    }
-
-    fn mounted_spell_store_with_no_aura_cancel(
-        spell_id: i32,
-        creature_entry: i32,
-    ) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            spell_id,
-            wow_data::SpellInfo {
-                spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                effect_base_points: 77,
-                effect_bonus_coefficient: 0.0,
-                aura_type: Some(wow_data::spell::aura_types::SPELL_AURA_MOUNTED),
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: vec![wow_data::SpellEffectInfo {
-                    effect_index: 0,
-                    effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                    effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOUNTED,
-                    effect_base_points: 77,
-                    effect_misc_value_1: creature_entry,
-                    ..Default::default()
-                }],
-            },
-        );
-        let mut attributes = [0; 15];
-        attributes[0] = wow_data::spell::attributes::SPELL_ATTR0_NO_AURA_CANCEL;
-        spell_store.insert_spell_misc_attributes_like_cpp(spell_id, attributes);
-        Arc::new(spell_store)
-    }
-
-    fn channeled_spell_store(spell_id: i32) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            spell_id,
-            wow_data::SpellInfo {
-                spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: 0,
-                effect_base_points: 0,
-                effect_bonus_coefficient: 0.0,
-                aura_type: None,
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: Vec::new(),
-            },
-        );
-        let mut attributes = [0; 15];
-        attributes[1] = wow_data::spell::attributes::SPELL_ATTR1_IS_CHANNELLED;
-        spell_store.insert_spell_misc_attributes_like_cpp(spell_id, attributes);
-        Arc::new(spell_store)
-    }
-
-    fn channeled_spell_store_with_no_aura_cancel(spell_id: i32) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            spell_id,
-            wow_data::SpellInfo {
-                spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: 0,
-                effect_base_points: 0,
-                effect_bonus_coefficient: 0.0,
-                aura_type: None,
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: Vec::new(),
-            },
-        );
-        let mut attributes = [0; 15];
-        attributes[0] = wow_data::spell::attributes::SPELL_ATTR0_NO_AURA_CANCEL;
-        attributes[1] = wow_data::spell::attributes::SPELL_ATTR1_IS_CHANNELLED;
-        spell_store.insert_spell_misc_attributes_like_cpp(spell_id, attributes);
-        Arc::new(spell_store)
-    }
-
-    fn mod_scale_spell_store(spell_id: i32, no_aura_cancel: bool) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            spell_id,
-            wow_data::SpellInfo {
-                spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                effect_base_points: 50,
-                effect_bonus_coefficient: 0.0,
-                aura_type: Some(wow_data::spell::aura_types::SPELL_AURA_MOD_SCALE),
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: vec![wow_data::SpellEffectInfo {
-                    effect_index: 0,
-                    effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                    effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOD_SCALE,
-                    effect_base_points: 50,
-                    ..Default::default()
-                }],
-            },
-        );
-        if no_aura_cancel {
-            let mut attributes = [0; 15];
-            attributes[0] = wow_data::spell::attributes::SPELL_ATTR0_NO_AURA_CANCEL;
-            spell_store.insert_spell_misc_attributes_like_cpp(spell_id, attributes);
-        }
-        Arc::new(spell_store)
-    }
-
-    fn mod_speed_no_control_spell_store(
-        spell_id: i32,
-        no_aura_cancel: bool,
-    ) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            spell_id,
-            wow_data::SpellInfo {
-                spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                effect_base_points: 50,
-                effect_bonus_coefficient: 0.0,
-                aura_type: Some(wow_data::spell::aura_types::SPELL_AURA_MOD_SPEED_NO_CONTROL),
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: vec![wow_data::SpellEffectInfo {
-                    effect_index: 0,
-                    effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-                    effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOD_SPEED_NO_CONTROL,
-                    effect_base_points: 50,
-                    ..Default::default()
-                }],
-            },
-        );
-        if no_aura_cancel {
-            let mut attributes = [0; 15];
-            attributes[0] = wow_data::spell::attributes::SPELL_ATTR0_NO_AURA_CANCEL;
-            spell_store.insert_spell_misc_attributes_like_cpp(spell_id, attributes);
-        }
-        Arc::new(spell_store)
-    }
-
-    fn drain_server_opcodes(send_rx: &flume::Receiver<Vec<u8>>) -> Vec<ServerOpcodes> {
-        let mut opcodes = Vec::new();
-        while let Ok(bytes) = send_rx.try_recv() {
-            if let Some(opcode) = WorldPacket::from_bytes(&bytes).server_opcode() {
-                opcodes.push(opcode);
-            }
-        }
-        opcodes
-    }
-
-    fn drain_server_packet_bytes(send_rx: &flume::Receiver<Vec<u8>>) -> Vec<Vec<u8>> {
-        let mut packets = Vec::new();
-        while let Ok(bytes) = send_rx.try_recv() {
-            packets.push(bytes);
-        }
-        packets
-    }
-
-    fn cast_failed_reason_like_cpp(bytes: &[u8]) -> i32 {
-        let mut packet = WorldPacket::from_bytes(bytes);
-        assert_eq!(
-            packet.server_opcode(),
-            Some(ServerOpcodes::CastFailed),
-            "expected CastFailed packet"
-        );
-        let _ = packet.read_uint16().expect("opcode");
-        let _ = packet.read_packed_guid().expect("cast id");
-        let _ = packet.read_int32().expect("spell id");
-        let _ = SpellCastVisual::read(&mut packet).expect("visual");
-        packet.read_int32().expect("reason")
-    }
-
-    fn cast_failed_fields_like_cpp(bytes: &[u8]) -> (ObjectGuid, i32, i32) {
-        let mut packet = WorldPacket::from_bytes(bytes);
-        assert_eq!(
-            packet.server_opcode(),
-            Some(ServerOpcodes::CastFailed),
-            "expected CastFailed packet"
-        );
-        let _ = packet.read_uint16().expect("opcode");
-        let cast_id = packet.read_packed_guid().expect("cast id");
-        let spell_id = packet.read_int32().expect("spell id");
-        let _ = SpellCastVisual::read(&mut packet).expect("visual");
-        let reason = packet.read_int32().expect("reason");
-        (cast_id, spell_id, reason)
-    }
-
-    fn spell_go_spell_id_like_cpp(bytes: &[u8]) -> i32 {
-        let mut packet = WorldPacket::from_bytes(bytes);
-        assert_eq!(
-            packet.server_opcode(),
-            Some(ServerOpcodes::SpellGo),
-            "expected SpellGo packet"
-        );
-        let _ = packet.read_uint16().expect("opcode");
-        let _ = packet.read_packed_guid().expect("caster");
-        let _ = packet.read_packed_guid().expect("caster unit");
-        let _ = packet.read_packed_guid().expect("cast id");
-        let _ = packet.read_packed_guid().expect("original cast id");
-        packet.read_int32().expect("spell id")
-    }
-
-    fn mount_result_like_cpp(bytes: &[u8]) -> i32 {
-        let mut packet = WorldPacket::from_bytes(bytes);
-        assert_eq!(
-            packet.server_opcode(),
-            Some(ServerOpcodes::MountResult),
-            "expected MountResult packet"
-        );
-        let _ = packet.read_uint16().expect("opcode");
-        packet.read_int32().expect("result")
-    }
-
-    fn cancel_aura_packet(spell_id: i32, caster_guid: ObjectGuid) -> WorldPacket {
-        let mut pkt = WorldPacket::new_empty();
-        pkt.write_int32(spell_id);
-        pkt.write_packed_guid(&caster_guid);
-        pkt.reset_read();
-        pkt
-    }
-
-    fn cancel_mod_speed_no_control_packet(target_guid: ObjectGuid) -> WorldPacket {
-        let mut pkt = WorldPacket::new_empty();
-        pkt.write_packed_guid(&target_guid);
-        pkt.reset_read();
-        pkt
-    }
-
-    fn int32_spell_packet(spell_id: i32) -> WorldPacket {
-        let mut pkt = WorldPacket::new_empty();
-        pkt.write_int32(spell_id);
-        pkt.reset_read();
-        pkt
-    }
-
-    fn self_res_spell_store(spell_id: i32) -> Arc<wow_data::SpellStore> {
-        let mut spell_store = wow_data::SpellStore::new();
-        spell_store.insert(
-            spell_id,
-            wow_data::SpellInfo {
-                spell_id,
-                cast_time_ms: 0,
-                cooldown_ms: 0,
-                recovery_time_ms: 0,
-                effect_type: 0,
-                effect_base_points: 0,
-                effect_bonus_coefficient: 0.0,
-                aura_type: None,
-                display_flags: 0,
-                requires_spell_focus: 0,
-                power_costs: Vec::new(),
-                effects: vec![wow_data::SpellEffectInfo {
-                    effect_index: 0,
-                    effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_SELF_RESURRECT,
-                    effect_base_points: -35,
-                    effect_misc_value_1: 77,
-                    ..Default::default()
-                }],
-            },
-        );
-        Arc::new(spell_store)
-    }
-
-    fn pet_cancel_aura_packet(pet_guid: ObjectGuid, spell_id: u32) -> WorldPacket {
-        let mut pkt = WorldPacket::new_empty();
-        pkt.write_packed_guid(&pet_guid);
-        pkt.write_uint32(spell_id);
-        pkt.reset_read();
-        pkt
-    }
-
-    fn totem_destroyed_packet(slot: u8, totem_guid: ObjectGuid) -> WorldPacket {
-        let mut pkt = WorldPacket::new_empty();
-        pkt.write_uint8(slot);
-        pkt.write_packed_guid(&totem_guid);
-        pkt.reset_read();
-        pkt
-    }
-
-    #[tokio::test]
-    async fn cancel_cast_clears_matching_active_cast_like_cpp() {
-        let (mut session, _send_rx) = make_session();
-        let cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 7);
-        install_active_spell_cast(&mut session, 12_345, cast_id);
-
-        session
-            .handle_cancel_cast(cancel_cast_packet(cast_id, 12_345))
-            .await;
-
-        assert!(session.active_spell_cast_snapshot_like_cpp().is_none());
-    }
-
-    #[tokio::test]
-    async fn cancel_cast_also_cancels_pending_spell_request_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let active_cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 7);
-        let pending_cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 8);
-        install_active_spell_cast(&mut session, 12_345, active_cast_id);
-        install_pending_spell_cast_request(&mut session, 67_890, pending_cast_id);
-
-        session
-            .handle_cancel_cast(cancel_cast_packet(active_cast_id, 12_345))
-            .await;
-
-        assert!(session.active_spell_cast_snapshot_like_cpp().is_none());
-        assert!(session.pending_spell_cast_for_test_like_cpp().is_none());
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 1);
-        assert_eq!(
-            cast_failed_fields_like_cpp(&packets[0]),
-            (pending_cast_id, 67_890, 32),
-            "C++ HandleCancelCastOpcode calls Player::CancelPendingCastRequest after interrupting"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_cast_mismatched_spell_preserves_active_cast_like_cpp() {
-        let (mut session, _send_rx) = make_session();
-        let cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 7);
-        install_active_spell_cast(&mut session, 12_345, cast_id);
-
-        session
-            .handle_cancel_cast(cancel_cast_packet(cast_id, 67_890))
-            .await;
-
-        assert_eq!(
-            session
-                .active_spell_cast_snapshot_like_cpp()
-                .as_ref()
-                .map(|active_cast| active_cast.spell_id),
-            Some(12_345)
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_cast_mismatch_cancels_pending_but_preserves_active_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let active_cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 7);
-        let pending_cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 8);
-        install_active_spell_cast(&mut session, 12_345, active_cast_id);
-        install_pending_spell_cast_request(&mut session, 67_890, pending_cast_id);
-
-        session
-            .handle_cancel_cast(cancel_cast_packet(active_cast_id, 54_321))
-            .await;
-
-        assert_eq!(
-            session
-                .active_spell_cast_snapshot_like_cpp()
-                .as_ref()
-                .map(|active_cast| active_cast.spell_id),
-            Some(12_345)
-        );
-        assert!(session.pending_spell_cast_for_test_like_cpp().is_none());
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 1);
-        assert_eq!(
-            cast_failed_fields_like_cpp(&packets[0]),
-            (pending_cast_id, 67_890, 32),
-            "Classic SpellHandler.cpp:263 cancels the pending request whenever a non-melee cast exists"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_channelling_interrupts_matching_player_channel_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 8);
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_spell_store(basic_spell_store([12_345]));
-        install_active_spell_cast(&mut session, 12_345, cast_id);
-        install_canonical_channeled_spell(&mut session, player_guid, 12_345);
-
-        session
-            .handle_cancel_channelling(cancel_channelling_packet(12_345, 40))
-            .await;
-
-        assert_eq!(canonical_channeled_spell_id(&mut session), None);
-        assert!(session.active_spell_cast_snapshot_like_cpp().is_none());
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancel_channelling_mismatched_spell_preserves_channel_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 9);
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_spell_store(basic_spell_store([67_890]));
-        install_active_spell_cast(&mut session, 12_345, cast_id);
-        let spell = install_canonical_channeled_spell(&mut session, player_guid, 12_345);
-
-        session
-            .handle_cancel_channelling(cancel_channelling_packet(67_890, 40))
-            .await;
-
-        assert_eq!(
-            canonical_channeled_spell_id(&mut session),
-            Some(spell.spell_id)
-        );
-        assert_eq!(
-            session
-                .active_spell_cast_snapshot_like_cpp()
-                .as_ref()
-                .map(|active_cast| active_cast.spell_id),
-            Some(12_345)
-        );
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancel_channelling_no_aura_cancel_spell_preserves_channel_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 12);
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_spell_store(mounted_spell_store_with_no_aura_cancel(12_345, 0));
-        install_active_spell_cast(&mut session, 12_345, cast_id);
-        let spell = install_canonical_channeled_spell(&mut session, player_guid, 12_345);
-
-        session
-            .handle_cancel_channelling(cancel_channelling_packet(12_345, 40))
-            .await;
-
-        assert_eq!(
-            canonical_channeled_spell_id(&mut session),
-            Some(spell.spell_id)
-        );
-        assert_eq!(
-            session
-                .active_spell_cast_snapshot_like_cpp()
-                .as_ref()
-                .map(|active_cast| active_cast.spell_id),
-            Some(12_345)
-        );
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancel_channelling_zero_spell_preserves_channel_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 10);
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_spell_store(basic_spell_store([12_345]));
-        install_active_spell_cast(&mut session, 12_345, cast_id);
-        let spell = install_canonical_channeled_spell(&mut session, player_guid, 12_345);
-
-        session
-            .handle_cancel_channelling(cancel_channelling_packet(0, 40))
-            .await;
-
-        assert_eq!(
-            canonical_channeled_spell_id(&mut session),
-            Some(spell.spell_id)
-        );
-        assert_eq!(
-            session
-                .active_spell_cast_snapshot_like_cpp()
-                .as_ref()
-                .map(|active_cast| active_cast.spell_id),
-            Some(12_345)
-        );
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancel_channelling_missing_spellinfo_preserves_channel_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 11);
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_spell_store(basic_spell_store([]));
-        install_active_spell_cast(&mut session, 12_345, cast_id);
-        let spell = install_canonical_channeled_spell(&mut session, player_guid, 12_345);
-
-        session
-            .handle_cancel_channelling(cancel_channelling_packet(12_345, 40))
-            .await;
-
-        assert_eq!(
-            canonical_channeled_spell_id(&mut session),
-            Some(spell.spell_id)
-        );
-        assert_eq!(
-            session
-                .active_spell_cast_snapshot_like_cpp()
-                .as_ref()
-                .map(|active_cast| active_cast.spell_id),
-            Some(12_345)
-        );
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancel_queued_spell_clears_only_pending_request_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let active_cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 7);
-        let pending_cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 8);
-        install_active_spell_cast(&mut session, 12_345, active_cast_id);
-        install_pending_spell_cast_request(&mut session, 67_890, pending_cast_id);
-
-        session
-            .handle_cancel_queued_spell(WorldPacket::new_empty())
-            .await;
-
-        assert_eq!(
-            session
-                .active_spell_cast_snapshot_like_cpp()
-                .as_ref()
-                .map(|active_cast| active_cast.spell_id),
-            Some(12_345)
-        );
-        assert!(session.pending_spell_cast_for_test_like_cpp().is_none());
-
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 1);
-        assert_eq!(
-            cast_failed_fields_like_cpp(&packets[0]),
-            (pending_cast_id, 67_890, 32),
-            "C++ Player::CancelPendingCastRequest sends SPELL_FAILED_DONT_REPORT for the queued request"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_queued_spell_without_pending_request_is_silent_like_cpp() {
-        let (mut session, send_rx) = make_session();
-
-        session
-            .handle_cancel_queued_spell(WorldPacket::new_empty())
-            .await;
-
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancel_aura_without_matching_represented_aura_stays_silent_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let caster_guid = ObjectGuid::create_player(1, 42);
-
-        session
-            .handle_cancel_aura(cancel_aura_packet(12_345, caster_guid))
-            .await;
-
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancel_aura_channeled_spell_interrupts_matching_channel_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 13);
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_spell_store(channeled_spell_store(12_345));
-        install_active_spell_cast(&mut session, 12_345, cast_id);
-        install_canonical_channeled_spell(&mut session, player_guid, 12_345);
-
-        session
-            .handle_cancel_aura(cancel_aura_packet(12_345, ObjectGuid::EMPTY))
-            .await;
-
-        assert_eq!(canonical_channeled_spell_id(&mut session), None);
-        assert!(session.active_spell_cast_snapshot_like_cpp().is_none());
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancel_aura_channeled_mismatched_current_spell_preserves_channel_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 14);
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_spell_store(channeled_spell_store(67_890));
-        install_active_spell_cast(&mut session, 12_345, cast_id);
-        let spell = install_canonical_channeled_spell(&mut session, player_guid, 12_345);
-
-        session
-            .handle_cancel_aura(cancel_aura_packet(67_890, ObjectGuid::EMPTY))
-            .await;
-
-        assert_eq!(
-            canonical_channeled_spell_id(&mut session),
-            Some(spell.spell_id)
-        );
-        assert_eq!(
-            session
-                .active_spell_cast_snapshot_like_cpp()
-                .as_ref()
-                .map(|active_cast| active_cast.spell_id),
-            Some(12_345)
-        );
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancel_aura_channeled_no_aura_cancel_preserves_channel_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 15);
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_spell_store(channeled_spell_store_with_no_aura_cancel(12_345));
-        install_active_spell_cast(&mut session, 12_345, cast_id);
-        let spell = install_canonical_channeled_spell(&mut session, player_guid, 12_345);
-
-        session
-            .handle_cancel_aura(cancel_aura_packet(12_345, ObjectGuid::EMPTY))
-            .await;
-
-        assert_eq!(
-            canonical_channeled_spell_id(&mut session),
-            Some(spell.spell_id)
-        );
-        assert_eq!(
-            session
-                .active_spell_cast_snapshot_like_cpp()
-                .as_ref()
-                .map(|active_cast| active_cast.spell_id),
-            Some(12_345)
-        );
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancel_aura_removes_matching_represented_mount_aura_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let caster_guid = ObjectGuid::create_player(1, 42);
-        session.set_player_guid(Some(caster_guid));
-        let effect = wow_data::SpellEffectInfo {
-            effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-            effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOUNTED,
-            effect_base_points: 77,
-            effect_misc_value_1: 0,
-            ..Default::default()
-        };
-
-        session
-            .apply_represented_mounted_aura_for_test_like_cpp(12_345, caster_guid, &effect)
-            .unwrap();
-        session.set_spell_store(mounted_spell_store(12_345, 0));
-        assert!(session.player_mounted_like_cpp());
-
-        session
-            .handle_cancel_aura(cancel_aura_packet(12_345, caster_guid))
-            .await;
-
-        assert!(!session.player_mounted_like_cpp());
-        assert!(!send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancel_aura_no_aura_cancel_spell_preserves_represented_mount_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let caster_guid = ObjectGuid::create_player(1, 42);
-        let effect = wow_data::SpellEffectInfo {
-            effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-            effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOUNTED,
-            effect_base_points: 77,
-            effect_misc_value_1: 0,
-            ..Default::default()
-        };
-
-        session
-            .apply_represented_mounted_aura_for_test_like_cpp(12_345, caster_guid, &effect)
-            .unwrap();
-        session.set_spell_store(mounted_spell_store_with_no_aura_cancel(12_345, 0));
-        let _ = drain_server_opcodes(&send_rx);
-
-        session
-            .handle_cancel_aura(cancel_aura_packet(12_345, caster_guid))
-            .await;
-
-        assert!(session.player_mounted_like_cpp());
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancel_aura_preserves_represented_mount_from_other_caster_like_cpp() {
-        let (mut session, _send_rx) = make_session();
-        let caster_guid = ObjectGuid::create_player(1, 42);
-        let other_caster_guid = ObjectGuid::create_player(1, 43);
-        let effect = wow_data::SpellEffectInfo {
-            effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-            effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOUNTED,
-            effect_base_points: 77,
-            effect_misc_value_1: 0,
-            ..Default::default()
-        };
-
-        session
-            .apply_represented_mounted_aura_for_test_like_cpp(12_345, caster_guid, &effect)
-            .unwrap();
-        session.set_spell_store(mounted_spell_store(12_345, 0));
-
-        session
-            .handle_cancel_aura(cancel_aura_packet(12_345, other_caster_guid))
-            .await;
-
-        assert!(session.player_mounted_like_cpp());
-    }
-
-    #[tokio::test]
-    async fn cancel_aura_empty_caster_matches_represented_mount_like_cpp() {
-        let (mut session, _send_rx) = make_session();
-        let caster_guid = ObjectGuid::create_player(1, 42);
-        let effect = wow_data::SpellEffectInfo {
-            effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-            effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOUNTED,
-            effect_base_points: 77,
-            effect_misc_value_1: 0,
-            ..Default::default()
-        };
-
-        session
-            .apply_represented_mounted_aura_for_test_like_cpp(12_345, caster_guid, &effect)
-            .unwrap();
-        session.set_spell_store(mounted_spell_store(12_345, 0));
-
-        session
-            .handle_cancel_aura(cancel_aura_packet(12_345, ObjectGuid::EMPTY))
-            .await;
-
-        assert!(!session.player_mounted_like_cpp());
-    }
-
-    #[tokio::test]
-    async fn cancel_empty_spell_handlers_stay_silent_without_runtime_slots_like_cpp() {
-        let (mut session, send_rx) = make_session();
-
-        session
-            .handle_cancel_auto_repeat_spell(WorldPacket::new_empty())
-            .await;
-        session
-            .handle_cancel_growth_aura(WorldPacket::new_empty())
-            .await;
-        session
-            .handle_cancel_mount_aura(WorldPacket::new_empty())
-            .await;
-
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cancel_growth_aura_removes_represented_mod_scale_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        session.set_player_guid(Some(player_guid));
-        session.set_spell_store(mod_scale_spell_store(12_345, false));
-
-        session
-            .execute_spell(12_345, player_guid)
-            .await
-            .expect("represented mod-scale aura should apply");
-        let _ = drain_server_opcodes(&send_rx);
-        assert!(session.visible_auras.values().any(|aura| {
-            aura.represented_effect == Some(RepresentedAuraEffectLikeCpp::ModScale)
-        }));
-
-        session
-            .handle_cancel_growth_aura(WorldPacket::new_empty())
-            .await;
-
-        assert!(!session.visible_auras.values().any(|aura| {
-            aura.represented_effect == Some(RepresentedAuraEffectLikeCpp::ModScale)
-        }));
-    }
-
-    #[tokio::test]
-    async fn cancel_growth_aura_no_aura_cancel_preserves_mod_scale_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        session.set_player_guid(Some(player_guid));
-        session.set_spell_store(mod_scale_spell_store(12_345, true));
-
-        session
-            .execute_spell(12_345, player_guid)
-            .await
-            .expect("represented no-aura-cancel mod-scale aura should apply");
-        let _ = drain_server_opcodes(&send_rx);
-
-        session
-            .handle_cancel_growth_aura(WorldPacket::new_empty())
-            .await;
-
-        assert!(session.visible_auras.values().any(|aura| {
-            aura.represented_effect == Some(RepresentedAuraEffectLikeCpp::ModScale)
-        }));
-    }
-
-    #[tokio::test]
-    async fn cancel_mod_speed_no_control_removes_matching_mover_aura_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        session.set_player_guid(Some(player_guid));
-        session.set_spell_store(mod_speed_no_control_spell_store(12_345, false));
-
-        session
-            .execute_spell(12_345, player_guid)
-            .await
-            .expect("represented mod-speed-no-control aura should apply");
-        let _ = drain_server_opcodes(&send_rx);
-        assert!(session.visible_auras.values().any(|aura| {
-            aura.represented_effect == Some(RepresentedAuraEffectLikeCpp::ModSpeedNoControl)
-        }));
-
-        assert!(
-            session
-                .try_handle_cancel_mod_speed_no_control_auras_like_cpp(
-                    cancel_mod_speed_no_control_packet(player_guid),
-                )
-                .await
-        );
-
-        assert!(!session.visible_auras.values().any(|aura| {
-            aura.represented_effect == Some(RepresentedAuraEffectLikeCpp::ModSpeedNoControl)
-        }));
-    }
-
-    #[tokio::test]
-    async fn cancel_mod_speed_no_control_ignores_non_mover_guid_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let other_guid = ObjectGuid::create_player(1, 43);
-        session.set_player_guid(Some(player_guid));
-        session.set_spell_store(mod_speed_no_control_spell_store(12_345, false));
-
-        session
-            .execute_spell(12_345, player_guid)
-            .await
-            .expect("represented mod-speed-no-control aura should apply");
-        let _ = drain_server_opcodes(&send_rx);
-
-        assert!(
-            !session
-                .try_handle_cancel_mod_speed_no_control_auras_like_cpp(
-                    cancel_mod_speed_no_control_packet(other_guid),
-                )
-                .await
-        );
-
-        assert!(session.visible_auras.values().any(|aura| {
-            aura.represented_effect == Some(RepresentedAuraEffectLikeCpp::ModSpeedNoControl)
-        }));
-    }
-
-    #[tokio::test]
-    async fn cancel_mod_speed_no_control_no_aura_cancel_preserves_aura_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        session.set_player_guid(Some(player_guid));
-        session.set_spell_store(mod_speed_no_control_spell_store(12_345, true));
-
-        session
-            .execute_spell(12_345, player_guid)
-            .await
-            .expect("represented no-aura-cancel mod-speed-no-control aura should apply");
-        let _ = drain_server_opcodes(&send_rx);
-
-        assert!(
-            session
-                .try_handle_cancel_mod_speed_no_control_auras_like_cpp(
-                    cancel_mod_speed_no_control_packet(player_guid),
-                )
-                .await
-        );
-
-        assert!(session.visible_auras.values().any(|aura| {
-            aura.represented_effect == Some(RepresentedAuraEffectLikeCpp::ModSpeedNoControl)
-        }));
-    }
-
-    #[tokio::test]
-    async fn cancel_mount_aura_no_aura_cancel_spell_preserves_mount_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let caster_guid = ObjectGuid::create_player(1, 42);
-        let effect = wow_data::SpellEffectInfo {
-            effect: wow_data::spell::spell_effect_types::SPELL_EFFECT_APPLY_AURA,
-            effect_aura: wow_data::spell::aura_types::SPELL_AURA_MOUNTED,
-            effect_base_points: 77,
-            effect_misc_value_1: 0,
-            ..Default::default()
-        };
-
-        session
-            .apply_represented_mounted_aura_for_test_like_cpp(12_345, caster_guid, &effect)
-            .unwrap();
-        session.set_spell_store(mounted_spell_store_with_no_aura_cancel(12_345, 0));
-        let _ = drain_server_opcodes(&send_rx);
-
-        session
-            .handle_cancel_mount_aura(WorldPacket::new_empty())
-            .await;
-
-        assert!(session.player_mounted_like_cpp());
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn self_res_unlisted_spell_stays_silent_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let spell_id = 20_000;
-
-        session.set_player_guid(Some(ObjectGuid::create_player(1, 20_000)));
-        session.set_spell_store(self_res_spell_store(spell_id));
-        session.set_player_health_like_cpp(0, 100);
-
-        session.handle_self_res(int32_spell_packet(spell_id)).await;
-
-        assert!(!session.player_is_alive_like_cpp());
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn self_res_listed_spell_casts_and_removes_self_res_spell_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let spell_id = 20_001;
-
-        session.set_player_guid(Some(ObjectGuid::create_player(1, 20_001)));
-        session.set_spell_store(self_res_spell_store(spell_id));
-        session.set_player_health_like_cpp(0, 100);
-        session.add_represented_self_res_spell_like_cpp(spell_id);
-
-        session.handle_self_res(int32_spell_packet(spell_id)).await;
-
-        assert!(session.player_is_alive_like_cpp());
-        assert_eq!(session.player_health_like_cpp(), 35);
-        assert!(!session.has_represented_self_res_spell_like_cpp(spell_id));
-        assert!(!send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn self_res_missing_spell_info_keeps_self_res_spell_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let spell_id = 20_002;
-
-        session.set_player_guid(Some(ObjectGuid::create_player(1, 20_002)));
-        session.set_player_health_like_cpp(0, 100);
-        session.add_represented_self_res_spell_like_cpp(spell_id);
-
-        session.handle_self_res(int32_spell_packet(spell_id)).await;
-
-        assert!(!session.player_is_alive_like_cpp());
-        assert!(session.has_represented_self_res_spell_like_cpp(spell_id));
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn pet_cancel_aura_removes_owned_pet_aura_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let pet_guid = ObjectGuid::create_world_object(HighGuid::Pet, 0, 1, 571, 0, 777, 42);
-        let spell_id = 12_345;
-        let aura = AppliedAuraRef::new(spell_id, ObjectGuid::EMPTY, 0, 0x1);
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_spell_store(basic_spell_store([spell_id as i32]));
-        set_canonical_player_pet_guid(&canonical, player_guid, pet_guid);
-        add_canonical_test_pet_on_map(&canonical, player_guid, pet_guid, spell_id, true);
-
-        session
-            .handle_pet_cancel_aura(pet_cancel_aura_packet(pet_guid, spell_id))
-            .await;
-
-        assert!(!canonical_pet_has_applied_aura(&canonical, pet_guid, aura));
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn cast_spell_applies_embedded_move_update_like_cpp() {
-        let (mut session, _send_rx) = make_session();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let spell_id = 13_337;
-        let moved_position = Position::new(33.0, 44.0, 55.0, 1.25);
-        let move_update = MovementInfo {
-            guid: player_guid,
-            time: 12_345,
-            position: moved_position,
-            ..MovementInfo::default()
-        };
-
-        session.set_player_guid(Some(player_guid));
-        session.set_player_map_position_like_cpp(571, Position::new(10.0, 20.0, 30.0, 0.0));
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_spell_store(basic_spell_store([spell_id]));
-
-        session
-            .handle_cast_spell(cast_spell_packet_with_move_update(
-                spell_id,
-                player_guid,
-                Some(move_update),
-            ))
-            .await;
-
-        assert_eq!(session.player_position_like_cpp(), Some(moved_position));
-    }
-
-    #[tokio::test]
-    async fn cast_spell_uses_represented_override_spell_info_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let original_spell_id = 13_347;
-        let override_spell_id = 13_348;
-
-        session.set_player_guid(Some(player_guid));
-        session.set_known_spells_like_cpp(vec![original_spell_id]);
-        session.set_spell_store(basic_spell_store([original_spell_id, override_spell_id]));
-        session.add_represented_override_spell_like_cpp(original_spell_id, override_spell_id);
-
-        session
-            .handle_cast_spell(cast_spell_packet(original_spell_id, player_guid))
-            .await;
-
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 4);
-        assert_eq!(
-            &packets[0][..2],
-            &(ServerOpcodes::SpellPrepare as u16).to_le_bytes()
-        );
-        assert_eq!(
-            &packets[1][..2],
-            &(ServerOpcodes::SpellStart as u16).to_le_bytes()
-        );
-        assert_eq!(
-            spell_go_spell_id_like_cpp(&packets[2]),
-            override_spell_id,
-            "C++ Player::GetCastSpellInfo resolves m_overrideSpells after the original spell known check"
-        );
-    }
-
-    #[tokio::test]
-    async fn cast_spell_falls_back_when_represented_override_missing_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let original_spell_id = 13_349;
-        let missing_override_spell_id = 13_350;
-
-        session.set_player_guid(Some(player_guid));
-        session.set_known_spells_like_cpp(vec![original_spell_id]);
-        session.set_spell_store(basic_spell_store([original_spell_id]));
-        session
-            .add_represented_override_spell_like_cpp(original_spell_id, missing_override_spell_id);
-
-        session
-            .handle_cast_spell(cast_spell_packet(original_spell_id, player_guid))
-            .await;
-
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 4);
-        assert_eq!(
-            &packets[0][..2],
-            &(ServerOpcodes::SpellPrepare as u16).to_le_bytes()
-        );
-        assert_eq!(
-            &packets[1][..2],
-            &(ServerOpcodes::SpellStart as u16).to_le_bytes()
-        );
-        assert_eq!(
-            spell_go_spell_id_like_cpp(&packets[2]),
-            original_spell_id,
-            "C++ Player::GetCastSpellInfo ignores override entries whose SpellInfo cannot be resolved"
-        );
-    }
-
-    #[tokio::test]
-    async fn cast_spell_with_mana_cost_deducts_power_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let spell_id = 13_351;
-        install_canonical_player(&mut session, &canonical, player_guid);
-        set_canonical_player_mana_like_cpp(&mut session, 500, 1000);
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_spell_store(spell_store_with_mana_power_cost_like_cpp(
-            spell_id, 50, 10.0,
-        ));
-
-        session
-            .handle_cast_spell(cast_spell_packet(spell_id, player_guid))
-            .await;
-
-        assert_eq!(
-            canonical_player_mana_like_cpp(&mut session),
-            350,
-            "C++ CalcPowerCost charges flat ManaCost plus PowerCostPct of create mana"
-        );
-        let opcodes = drain_server_opcodes(&send_rx);
-        assert!(
-            !opcodes.contains(&ServerOpcodes::CastFailed),
-            "accepted casts must not send SPELL_FAILED_NO_POWER"
-        );
-        assert!(
-            opcodes.contains(&ServerOpcodes::UpdateObject),
-            "accepted casts with represented power cost must send a player values update"
-        );
-    }
-
-    #[tokio::test]
-    async fn cast_spell_power_spend_survives_canonical_resync_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let spell_id = 13_356;
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_map_store(Arc::new(wow_data::MapStore::from_entries([
-            wow_data::MapEntry {
-                id: 571,
-                instance_type: wow_data::map::MAP_COMMON,
-                expansion_id: 0,
-                parent_map_id: -1,
-                cosmetic_parent_map_id: -1,
-                flags1: 0,
-                flags2: 0,
-            },
-        ])));
-        session.set_loaded_player_identity_like_cpp(571, 1, 5, 80, 0);
-        session.set_loaded_player_powers_like_cpp([500, 222, 0, 0, 0, 0, 0, 0, 0, 0]);
-        assert_eq!(
-            session.represented_player_power_values_like_cpp().unwrap()[1],
-            222
-        );
-        assert!(session.sync_canonical_player_primary_power_like_cpp(
-            PowerType::Mana,
-            500,
-            1_000,
-            1_000,
-        ));
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_spell_store(spell_store_with_mana_power_cost_like_cpp(
-            spell_id, 50, 10.0,
-        ));
-
-        session
-            .handle_cast_spell(cast_spell_packet(spell_id, player_guid))
-            .await;
-
-        assert_eq!(canonical_player_mana_like_cpp(&mut session), 350);
-        assert_eq!(
-            session.represented_player_power_values_like_cpp().unwrap()[0],
-            350,
-            "represented session power must mirror Spell::TakePower before later AddToMap-style resync"
-        );
-        assert_eq!(
-            session.represented_player_power_values_like_cpp().unwrap()[1],
-            222
-        );
-        let _ = session.ensure_canonical_world_map_for_current_player_like_cpp();
-        assert_eq!(
-            canonical_player_mana_like_cpp(&mut session),
-            350,
-            "resync from the session snapshot must not resurrect pre-cast mana"
-        );
-        assert_eq!(
-            session.represented_player_power_values_like_cpp().unwrap()[1],
-            222
-        );
-        let snapshot = session
-            .current_player_save_to_db_snapshot_like_cpp()
-            .unwrap();
-        assert_eq!(snapshot.powers[0], Some(350));
-        assert_eq!(
-            snapshot.powers[1],
-            Some(222),
-            "saving power1 after a cast must preserve other loaded character power columns"
-        );
-        let _ = drain_server_opcodes(&send_rx);
-    }
-
-    #[tokio::test]
-    async fn cast_spell_with_insufficient_mana_fails_no_power_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let spell_id = 13_352;
-        install_canonical_player(&mut session, &canonical, player_guid);
-        set_canonical_player_mana_like_cpp(&mut session, 149, 1000);
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_spell_store(spell_store_with_mana_power_cost_like_cpp(
-            spell_id, 50, 10.0,
-        ));
-
-        session
-            .handle_cast_spell(cast_spell_packet(spell_id, player_guid))
-            .await;
-
-        assert_eq!(
-            canonical_player_mana_like_cpp(&mut session),
-            149,
-            "failed casts must not deduct power"
-        );
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 2);
-        assert_eq!(
-            &packets[0][..2],
-            &(ServerOpcodes::SpellPrepare as u16).to_le_bytes()
-        );
-        assert_eq!(
-            cast_failed_reason_like_cpp(&packets[1]),
-            SpellCastResult::NoPower as i32
-        );
-    }
-
-    #[tokio::test]
-    async fn cast_spell_late_check_failure_does_not_deduct_power_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let spell_id = 13_353;
-        install_canonical_player(&mut session, &canonical, player_guid);
-        set_canonical_player_mana_like_cpp(&mut session, 500, 1000);
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_spell_store(spell_store_with_mana_cost_and_missing_focus_like_cpp(
-            spell_id, 150,
-        ));
-
-        session
-            .handle_cast_spell(cast_spell_packet(spell_id, player_guid))
-            .await;
-
-        assert_eq!(
-            canonical_player_mana_like_cpp(&mut session),
-            500,
-            "C++ Spell::TakePower happens after represented CheckCast failures such as missing spell focus"
-        );
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 2);
-        assert_eq!(
-            &packets[0][..2],
-            &(ServerOpcodes::SpellPrepare as u16).to_le_bytes()
-        );
-        assert_eq!(
-            cast_failed_reason_like_cpp(&packets[1]),
-            SpellCastResult::RequiresSpellFocus as i32
-        );
-    }
-
-    #[tokio::test]
-    async fn retained_direct_cast_late_power_failure_restores_legacy_timestamp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 22);
-        let spell_id = 13_357;
-        install_canonical_player(&mut session, &canonical, player_guid);
-        set_canonical_player_mana_like_cpp(&mut session, 149, 1000);
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_spell_store(spell_store_with_mana_power_cost_like_cpp(
-            spell_id, 50, 10.0,
-        ));
-        let previous_last_spell_cast_time =
-            Some(std::time::Instant::now() - std::time::Duration::from_millis(5_000));
-        session.mutate_cast_execution_like_cpp(|state| {
-            state.last_cast_time = previous_last_spell_cast_time
-        });
-        install_active_spell_cast(&mut session, spell_id, cast_id);
-        session.mutate_cast_execution_like_cpp(|state| {
-            if let Some(active) = state.active.as_mut() {
-                active.cast_start_time =
-                    std::time::Instant::now() - std::time::Duration::from_millis(30_000);
-                active.metadata = SpellCastMetadata {
-                    from_client: true,
-                    original_cast_id: cast_id,
-                    ..SpellCastMetadata::default()
-                };
-            }
-        });
-
-        session.tick_active_spell_cast().await;
-
-        assert!(session.active_spell_cast_snapshot_like_cpp().is_none());
-        assert_eq!(canonical_player_mana_like_cpp(&mut session), 149);
-        assert_eq!(
-            session.last_spell_cast_time_like_cpp().flatten(),
-            previous_last_spell_cast_time,
-            "Shared-executor consumers retain their legacy timestamp rollback contract; normal client preparation has separate GCD coverage"
-        );
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 1);
-        assert_eq!(
-            cast_failed_reason_like_cpp(&packets[0]),
-            SpellCastResult::NoPower as i32
-        );
-    }
-
-    #[tokio::test]
-    async fn queued_spell_with_mana_cost_deducts_when_executed_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let spell_id = 13_354;
-        install_canonical_player(&mut session, &canonical, player_guid);
-        set_canonical_player_mana_like_cpp(&mut session, 500, 1000);
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_spell_store(
-            spell_store_with_global_cooldown_and_mana_power_cost_like_cpp(
-                spell_id, 1_500, 50, 10.0,
-            ),
-        );
-        session.mutate_cast_execution_like_cpp(|state| {
-            state.last_cast_time =
-                Some(std::time::Instant::now() - std::time::Duration::from_millis(1_200))
-        });
-
-        session
-            .handle_cast_spell(cast_spell_packet(spell_id, player_guid))
-            .await;
-
-        assert!(
-            session
-                .pending_spell_cast_for_test_like_cpp()
-                .as_ref()
-                .is_some_and(|pending| pending.spell_id == spell_id)
-        );
-        assert!(
-            send_rx.is_empty(),
-            "C++ Player::_pendingSpellCastRequest queues without executing or spending power"
-        );
-        assert_eq!(canonical_player_mana_like_cpp(&mut session), 500);
-
-        session.mutate_cast_execution_like_cpp(|state| {
-            state.last_cast_time =
-                Some(std::time::Instant::now() - std::time::Duration::from_millis(1_500))
-        });
-        session.tick_pending_spell_cast_request_like_cpp().await;
-
-        assert!(session.pending_spell_cast_for_test_like_cpp().is_none());
-        assert_eq!(
-            canonical_player_mana_like_cpp(&mut session),
-            350,
-            "queued C++ casts still run Spell::TakePower when the pending request executes"
-        );
-        let opcodes = drain_server_opcodes(&send_rx);
-        assert!(opcodes.contains(&ServerOpcodes::UpdateObject));
-        assert!(opcodes.contains(&ServerOpcodes::SpellGo));
-        assert!(opcodes.contains(&ServerOpcodes::CooldownEvent));
-    }
-
-    #[tokio::test]
-    async fn mana_cost_spell_on_non_mana_class_does_not_spend_primary_power_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let spell_id = 13_355;
-        install_canonical_player(&mut session, &canonical, player_guid);
-        assert!(session.sync_canonical_player_primary_power_like_cpp(
-            PowerType::Rage,
-            500,
-            1000,
-            0
-        ));
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_spell_store(spell_store_with_mana_power_cost_like_cpp(spell_id, 50, 0.0));
-
-        session
-            .handle_cast_spell(cast_spell_packet(spell_id, player_guid))
-            .await;
-
-        assert_eq!(
-            canonical_player_power_like_cpp(&mut session, PowerType::Rage),
-            500,
-            "C++ GetPowerIndexByClass has no Mana slot for warrior, so mana-cost spells cannot spend Rage as Mana"
-        );
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 2);
-        assert_eq!(
-            &packets[0][..2],
-            &(ServerOpcodes::SpellPrepare as u16).to_le_bytes()
-        );
-        assert_eq!(
-            cast_failed_reason_like_cpp(&packets[1]),
-            SpellCastResult::NoPower as i32
-        );
-    }
-
-    #[tokio::test]
-    async fn cast_spell_rejects_gcd_outside_spell_queue_window_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let spell_id = 13_338;
-        session.set_player_guid(Some(player_guid));
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_spell_store(spell_store_with_global_cooldown(spell_id, 1_500));
-        session.mutate_cast_execution_like_cpp(|state| {
-            state.last_cast_time = Some(std::time::Instant::now())
-        });
-
-        session
-            .handle_cast_spell(cast_spell_packet(spell_id, player_guid))
-            .await;
-
-        assert!(session.pending_spell_cast_for_test_like_cpp().is_none());
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 1);
-        assert_eq!(
-            cast_failed_reason_like_cpp(&packets[0]),
-            SpellCastResult::SpellInProgress as i32,
-            "C++ HandleCastSpellOpcode sends SPELL_FAILED_SPELL_IN_PROGRESS when CanRequestSpellCast rejects the request"
-        );
-    }
-
-    #[tokio::test]
-    async fn cast_spell_queues_within_spell_queue_window_and_executes_after_gcd_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let spell_id = 13_339;
-        session.set_player_guid(Some(player_guid));
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_spell_store(spell_store_with_global_cooldown(spell_id, 1_500));
-        session.mutate_cast_execution_like_cpp(|state| {
-            state.last_cast_time =
-                Some(std::time::Instant::now() - std::time::Duration::from_millis(1_200))
-        });
-
-        session
-            .handle_cast_spell(cast_spell_packet(spell_id, player_guid))
-            .await;
-
-        assert!(
-            session
-                .pending_spell_cast_for_test_like_cpp()
-                .as_ref()
-                .is_some_and(|pending| pending.spell_id == spell_id)
-        );
-        assert!(
-            send_rx.is_empty(),
-            "C++ RequestSpellCast only queues while GCD is still active"
-        );
-
-        session.mutate_cast_execution_like_cpp(|state| {
-            state.last_cast_time =
-                Some(std::time::Instant::now() - std::time::Duration::from_millis(1_500))
-        });
-        session.tick_pending_spell_cast_request_like_cpp().await;
-
-        assert!(session.pending_spell_cast_for_test_like_cpp().is_none());
-        let opcodes = drain_server_opcodes(&send_rx);
-        assert!(opcodes.contains(&ServerOpcodes::SpellGo));
-        assert!(opcodes.contains(&ServerOpcodes::CooldownEvent));
-    }
-
-    #[tokio::test]
-    async fn cast_spell_rejects_active_cast_outside_spell_queue_window_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let active_cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 20);
-        let queued_spell_id = 13_340;
-        session.set_player_guid(Some(player_guid));
-        session.set_known_spells_like_cpp(vec![12_345, queued_spell_id]);
-        session.set_spell_store(basic_spell_store([12_345, queued_spell_id]));
-        install_active_spell_cast(&mut session, 12_345, active_cast_id);
-
-        session
-            .handle_cast_spell(cast_spell_packet(queued_spell_id, player_guid))
-            .await;
-
-        assert!(session.pending_spell_cast_for_test_like_cpp().is_none());
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 1);
-        assert_eq!(
-            cast_failed_reason_like_cpp(&packets[0]),
-            SpellCastResult::SpellInProgress as i32
-        );
-    }
-
-    #[tokio::test]
-    async fn cast_spell_queues_near_active_cast_finish_and_executes_after_cast_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let active_cast_id = ObjectGuid::create_world_object(HighGuid::Cast, 0, 1, 0, 0, 1, 21);
-        let queued_spell_id = 13_341;
-        session.set_player_guid(Some(player_guid));
-        session.set_known_spells_like_cpp(vec![12_345, queued_spell_id]);
-        session.set_spell_store(basic_spell_store([12_345, queued_spell_id]));
-        install_active_spell_cast(&mut session, 12_345, active_cast_id);
-        session.mutate_cast_execution_like_cpp(|state| {
-            if let Some(active) = state.active.as_mut() {
-                active.cast_start_time =
-                    std::time::Instant::now() - std::time::Duration::from_millis(29_700);
-            }
-        });
-
-        session
-            .handle_cast_spell(cast_spell_packet(queued_spell_id, player_guid))
-            .await;
-
-        assert!(
-            session
-                .pending_spell_cast_for_test_like_cpp()
-                .as_ref()
-                .is_some_and(|pending| pending.spell_id == queued_spell_id)
-        );
-        assert!(send_rx.is_empty());
-
-        session.mutate_cast_execution_like_cpp(|state| {
-            if let Some(active) = state.active.as_mut() {
-                active.cast_start_time =
-                    std::time::Instant::now() - std::time::Duration::from_millis(30_000);
-            }
-        });
-        session.tick_active_spell_cast().await;
-        session.tick_pending_spell_cast_request_like_cpp().await;
-
-        assert!(session.pending_spell_cast_for_test_like_cpp().is_none());
-        let opcodes = drain_server_opcodes(&send_rx);
-        assert!(opcodes.contains(&ServerOpcodes::SpellGo));
-        assert!(opcodes.contains(&ServerOpcodes::CooldownEvent));
-    }
-
-    #[tokio::test]
-    async fn cast_known_account_mount_spell_applies_mounted_aura_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let spell_id = 12345;
-
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_spell_store(mounted_spell_store(spell_id, 0));
-        session.set_mount_store(Arc::new(wow_data::MountStore::from_entries([
-            wow_data::MountEntry {
-                id: 7,
-                mount_type_id: 0,
-                flags: 0,
-                source_type_enum: 0,
-                source_spell_id: spell_id,
-                player_condition_id: 0,
-                mount_fly_ride_height: 0.0,
-                ui_model_scene_id: 0,
-            },
-        ])));
-        session.set_mount_x_display_store(Arc::new(wow_data::MountXDisplayStore::from_entries([
-            wow_data::MountXDisplayEntry {
-                id: 1,
-                creature_display_info_id: 4321,
-                player_condition_id: 0,
-                mount_id: 7,
-            },
-        ])));
-
-        session
-            .handle_cast_spell(cast_spell_packet(spell_id, player_guid))
-            .await;
-
-        assert!(session.player_mounted_like_cpp());
-        let opcodes = drain_server_opcodes(&send_rx);
-        assert!(opcodes.contains(&ServerOpcodes::SpellGo));
-        assert!(opcodes.contains(&ServerOpcodes::AuraUpdate));
-        assert!(opcodes.contains(&ServerOpcodes::UpdateObject));
-
-        let manager = canonical.lock().unwrap();
-        let player = manager
-            .find_map(571, 0)
-            .unwrap()
-            .map()
-            .get_typed_player(player_guid)
-            .unwrap();
-        assert_eq!(player.unit().data().mount_display_id, 4321);
-    }
-
-    #[tokio::test]
-    async fn cast_mount_spell_fails_not_here_without_mount_capability_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 43);
-        let spell_id = 12_346;
-
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_spell_store(mounted_spell_store(spell_id, 1234));
-        session.set_mount_store(Arc::new(wow_data::MountStore::from_entries([
-            wow_data::MountEntry {
-                id: 8,
-                mount_type_id: 7,
-                flags: 0,
-                source_type_enum: 0,
-                source_spell_id: spell_id,
-                player_condition_id: 0,
-                mount_fly_ride_height: 0.0,
-                ui_model_scene_id: 0,
-            },
-        ])));
-
-        session
-            .handle_cast_spell(cast_spell_packet(spell_id, player_guid))
-            .await;
-
-        assert!(!session.player_mounted_like_cpp());
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 2);
-        assert_eq!(
-            &packets[0][..2],
-            &(ServerOpcodes::SpellPrepare as u16).to_le_bytes()
-        );
-        assert_eq!(
-            cast_failed_reason_like_cpp(&packets[1]),
-            SpellCastResult::NotHere as i32
-        );
-    }
-
-    #[tokio::test]
-    async fn cast_flying_mount_spell_fails_only_abovewater_in_water_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 44);
-        let spell_id = 12_347;
-
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_player_liquid_status_like_cpp(crate::session::LIQUID_MAP_IN_WATER_LIKE_CPP);
-        session.set_spell_store(mounted_flying_spell_store(spell_id, 1234));
-        session.set_mount_store(Arc::new(wow_data::MountStore::from_entries([
-            wow_data::MountEntry {
-                id: 9,
-                mount_type_id: 0,
-                flags: 0,
-                source_type_enum: 0,
-                source_spell_id: spell_id,
-                player_condition_id: 0,
-                mount_fly_ride_height: 0.0,
-                ui_model_scene_id: 0,
-            },
-        ])));
-
-        session
-            .handle_cast_spell(cast_spell_packet(spell_id, player_guid))
-            .await;
-
-        assert!(!session.player_mounted_like_cpp());
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 2);
-        assert_eq!(
-            &packets[0][..2],
-            &(ServerOpcodes::SpellPrepare as u16).to_le_bytes()
-        );
-        assert_eq!(
-            cast_failed_reason_like_cpp(&packets[1]),
-            SpellCastResult::OnlyAbovewater as i32
-        );
-    }
-
-    #[tokio::test]
-    async fn cast_shapeshift_mount_form_fails_not_here_without_capability_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 45);
-        let spell_id = 12_348;
-        let form_id = 55;
-
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_known_spells_like_cpp(vec![spell_id]);
-        session.set_spell_store(shapeshift_spell_store(spell_id, form_id));
-        session.set_spell_shapeshift_form_store(Arc::new(
-            wow_data::SpellShapeshiftFormStore::from_entries([
-                wow_data::SpellShapeshiftFormEntry {
-                    id: form_id as u32,
-                    name: "Mounted Form".to_string(),
-                    creature_type: 0,
-                    flags: 0,
-                    attack_icon_file_id: 0,
-                    bonus_action_bar: 0,
-                    combat_round_time: 0,
-                    damage_variance: 0.0,
-                    mount_type_id: 7,
-                    creature_display_id: [0; 4],
-                    preset_spell_id: [0; wow_data::MAX_SHAPESHIFT_SPELLS],
-                },
-            ]),
-        ));
-
-        session
-            .handle_cast_spell(cast_spell_packet(spell_id, player_guid))
-            .await;
-
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 2);
-        assert_eq!(
-            &packets[0][..2],
-            &(ServerOpcodes::SpellPrepare as u16).to_le_bytes()
-        );
-        assert_eq!(
-            cast_failed_reason_like_cpp(&packets[1]),
-            SpellCastResult::NotHere as i32
-        );
-    }
-
-    #[tokio::test]
-    async fn cast_mount_spell_in_disallowed_shapeshift_form_sends_mount_result_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 46);
-        let mount_spell_id = 12_349;
-        let shapeshift_spell_id = 22_349;
-        let form_id = 56;
-
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_known_spells_like_cpp(vec![mount_spell_id]);
-        session.set_spell_store(mounted_spell_store_with_active_shapeshift_aura(
-            mount_spell_id,
-            shapeshift_spell_id,
-            form_id,
-        ));
-        session.set_spell_shapeshift_form_store(Arc::new(
-            wow_data::SpellShapeshiftFormStore::from_entries([
-                wow_data::SpellShapeshiftFormEntry {
-                    id: form_id as u32,
-                    name: "Non Stance Form".to_string(),
-                    creature_type: 0,
-                    flags: 0,
-                    attack_icon_file_id: 0,
-                    bonus_action_bar: 0,
-                    combat_round_time: 0,
-                    damage_variance: 0.0,
-                    mount_type_id: 0,
-                    creature_display_id: [0; 4],
-                    preset_spell_id: [0; wow_data::MAX_SHAPESHIFT_SPELLS],
-                },
-            ]),
-        ));
-        assert!(
-            session.insert_player_visible_aura_like_cpp(active_shapeshift_aura_for_test(
-                shapeshift_spell_id,
-                player_guid
-            ),)
-        );
-
-        session
-            .handle_cast_spell(cast_spell_packet(mount_spell_id, player_guid))
-            .await;
-
-        assert!(!session.player_mounted_like_cpp());
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 2);
-        assert_eq!(
-            &packets[0][..2],
-            &(ServerOpcodes::SpellPrepare as u16).to_le_bytes()
-        );
-        assert_eq!(mount_result_like_cpp(&packets[1]), 8);
-    }
-
-    #[tokio::test]
-    async fn cast_mount_spell_in_stance_shapeshift_form_is_allowed_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 47);
-        let mount_spell_id = 12_350;
-        let shapeshift_spell_id = 22_350;
-        let form_id = 57;
-
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_known_spells_like_cpp(vec![mount_spell_id]);
-        session.set_spell_store(mounted_spell_store_with_active_shapeshift_aura(
-            mount_spell_id,
-            shapeshift_spell_id,
-            form_id,
-        ));
-        session.set_spell_shapeshift_form_store(Arc::new(
-            wow_data::SpellShapeshiftFormStore::from_entries([
-                wow_data::SpellShapeshiftFormEntry {
-                    id: form_id as u32,
-                    name: "Stance Form".to_string(),
-                    creature_type: 0,
-                    flags: 0x0000_0001,
-                    attack_icon_file_id: 0,
-                    bonus_action_bar: 0,
-                    combat_round_time: 0,
-                    damage_variance: 0.0,
-                    mount_type_id: 0,
-                    creature_display_id: [0; 4],
-                    preset_spell_id: [0; wow_data::MAX_SHAPESHIFT_SPELLS],
-                },
-            ]),
-        ));
-        assert!(
-            session.insert_player_visible_aura_like_cpp(active_shapeshift_aura_for_test(
-                shapeshift_spell_id,
-                player_guid
-            ),)
-        );
-
-        session
-            .handle_cast_spell(cast_spell_packet(mount_spell_id, player_guid))
-            .await;
-
-        assert!(session.player_mounted_like_cpp());
-        let opcodes = drain_server_opcodes(&send_rx);
-        assert!(!opcodes.contains(&ServerOpcodes::MountResult));
-        assert!(!opcodes.contains(&ServerOpcodes::CastFailed));
-        assert!(opcodes.contains(&ServerOpcodes::SpellGo));
-    }
-
-    #[tokio::test]
-    async fn cast_mount_spell_in_disallowed_transformed_display_sends_mount_result_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 48);
-        let mount_spell_id = 12_351;
-        let transformed_display_id = 88_001;
-
-        install_canonical_player(&mut session, &canonical, player_guid);
-        set_canonical_player_display_for_test(
-            &canonical,
-            player_guid,
-            transformed_display_id,
-            false,
-        );
-        session.set_known_spells_like_cpp(vec![mount_spell_id]);
-        session.set_spell_store(mounted_spell_store(mount_spell_id, 0));
-        set_transformed_display_mount_check_stores_for_test(
-            &mut session,
-            transformed_display_id,
-            0,
-            0,
-        );
-
-        session
-            .handle_cast_spell(cast_spell_packet(mount_spell_id, player_guid))
-            .await;
-
-        assert!(!session.player_mounted_like_cpp());
-        let packets = drain_server_packet_bytes(&send_rx);
-        assert_eq!(packets.len(), 2);
-        assert_eq!(
-            &packets[0][..2],
-            &(ServerOpcodes::SpellPrepare as u16).to_le_bytes()
-        );
-        assert_eq!(mount_result_like_cpp(&packets[1]), 8);
-    }
-
-    #[tokio::test]
-    async fn cast_mount_spell_with_mountable_transform_spell_is_allowed_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 50);
-        let mount_spell_id = 12_353;
-        let transform_spell_id = 22_353;
-        let transformed_display_id = 88_003;
-
-        install_canonical_player(&mut session, &canonical, player_guid);
-        set_canonical_player_display_for_test(
-            &canonical,
-            player_guid,
-            transformed_display_id,
-            false,
-        );
-        session.set_known_spells_like_cpp(vec![mount_spell_id]);
-        session.set_spell_store(mounted_spell_store_with_transform_spell(
-            mount_spell_id,
-            transform_spell_id,
-            true,
-        ));
-        set_transformed_display_mount_check_stores_for_test(
-            &mut session,
-            transformed_display_id,
-            0,
-            0,
-        );
-        assert!(
-            session.insert_player_visible_aura_like_cpp(active_shapeshift_aura_for_test(
-                transform_spell_id,
-                player_guid
-            ),)
-        );
-
-        session
-            .handle_cast_spell(cast_spell_packet(mount_spell_id, player_guid))
-            .await;
-
-        assert!(session.player_mounted_like_cpp());
-        let opcodes = drain_server_opcodes(&send_rx);
-        assert!(!opcodes.contains(&ServerOpcodes::MountResult));
-        assert!(!opcodes.contains(&ServerOpcodes::CastFailed));
-        assert!(opcodes.contains(&ServerOpcodes::SpellGo));
-    }
-
-    #[tokio::test]
-    async fn cast_mount_spell_in_mountable_transformed_model_is_allowed_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 49);
-        let mount_spell_id = 12_352;
-        let transformed_display_id = 88_002;
-
-        install_canonical_player(&mut session, &canonical, player_guid);
-        set_canonical_player_display_for_test(
-            &canonical,
-            player_guid,
-            transformed_display_id,
-            false,
-        );
-        session.set_known_spells_like_cpp(vec![mount_spell_id]);
-        session.set_spell_store(mounted_spell_store(mount_spell_id, 0));
-        set_transformed_display_mount_check_stores_for_test(
-            &mut session,
-            transformed_display_id,
-            0x0000_0080,
-            0,
-        );
-
-        session
-            .handle_cast_spell(cast_spell_packet(mount_spell_id, player_guid))
-            .await;
-
-        assert!(session.player_mounted_like_cpp());
-        let opcodes = drain_server_opcodes(&send_rx);
-        assert!(!opcodes.contains(&ServerOpcodes::MountResult));
-        assert!(!opcodes.contains(&ServerOpcodes::CastFailed));
-        assert!(opcodes.contains(&ServerOpcodes::SpellGo));
-    }
-
-    #[tokio::test]
-    async fn pet_cancel_aura_missing_spellinfo_preserves_pet_aura_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let pet_guid = ObjectGuid::create_world_object(HighGuid::Pet, 0, 1, 571, 0, 777, 43);
-        let spell_id = 12_346;
-        let aura = AppliedAuraRef::new(spell_id, ObjectGuid::EMPTY, 0, 0x1);
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_spell_store(basic_spell_store([]));
-        set_canonical_player_pet_guid(&canonical, player_guid, pet_guid);
-        add_canonical_test_pet_on_map(&canonical, player_guid, pet_guid, spell_id, true);
-
-        session
-            .handle_pet_cancel_aura(pet_cancel_aura_packet(pet_guid, spell_id))
-            .await;
-
-        assert!(canonical_pet_has_applied_aura(&canonical, pet_guid, aura));
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn pet_cancel_aura_non_owned_pet_preserves_aura_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let other_player_guid = ObjectGuid::create_player(1, 43);
-        let pet_guid = ObjectGuid::create_world_object(HighGuid::Pet, 0, 1, 571, 0, 777, 44);
-        let spell_id = 12_347;
-        let aura = AppliedAuraRef::new(spell_id, ObjectGuid::EMPTY, 0, 0x1);
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_spell_store(basic_spell_store([spell_id as i32]));
-        add_canonical_test_pet_on_map(&canonical, other_player_guid, pet_guid, spell_id, true);
-
-        session
-            .handle_pet_cancel_aura(pet_cancel_aura_packet(pet_guid, spell_id))
-            .await;
-
-        assert!(canonical_pet_has_applied_aura(&canonical, pet_guid, aura));
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn pet_cancel_aura_dead_pet_sends_feedback_and_preserves_aura_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let pet_guid = ObjectGuid::create_world_object(HighGuid::Pet, 0, 1, 571, 0, 777, 45);
-        let spell_id = 12_348;
-        let aura = AppliedAuraRef::new(spell_id, ObjectGuid::EMPTY, 0, 0x1);
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_spell_store(basic_spell_store([spell_id as i32]));
-        set_canonical_player_pet_guid(&canonical, player_guid, pet_guid);
-        add_canonical_test_pet_on_map(&canonical, player_guid, pet_guid, spell_id, false);
-
-        session
-            .handle_pet_cancel_aura(pet_cancel_aura_packet(pet_guid, spell_id))
-            .await;
-
-        assert!(canonical_pet_has_applied_aura(&canonical, pet_guid, aura));
-        let bytes = send_rx.try_recv().expect("pet action feedback");
-        assert_eq!(
-            u16::from_le_bytes([bytes[0], bytes[1]]),
-            ServerOpcodes::PetActionFeedback as u16
-        );
-        assert_eq!(&bytes[2..6], &0i32.to_le_bytes());
-        assert_eq!(
-            bytes[6],
-            wow_packet::packets::pet::PET_ACTION_FEEDBACK_DEAD_LIKE_CPP
-        );
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn pet_cancel_aura_removes_charmed_creature_aura_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let creature_guid =
-            ObjectGuid::create_world_object(HighGuid::Creature, 0, 1, 571, 0, 777, 46);
-        let spell_id = 12_349;
-        let aura = AppliedAuraRef::new(spell_id, ObjectGuid::EMPTY, 0, 0x1);
-        install_canonical_player(&mut session, &canonical, player_guid);
-        session.set_spell_store(basic_spell_store([spell_id as i32]));
-        set_canonical_player_charmed_guid(&canonical, player_guid, creature_guid);
-        add_canonical_test_creature_on_map(
-            &canonical,
-            creature_guid,
-            Position::new(11.0, 21.0, 30.0, 0.0),
-            571,
-            0,
-            false,
-        );
-        {
-            let mut guard = canonical.lock().unwrap();
-            let creature = guard
-                .find_map_mut(571, 0)
-                .unwrap()
-                .map_mut()
-                .get_typed_creature_mut(creature_guid)
-                .unwrap();
-            creature.unit_mut().set_max_health(100);
-            creature.unit_mut().set_health(100);
-            creature.unit_mut().subsystems_mut().auras.add_applied(aura);
-        }
-
-        session
-            .handle_pet_cancel_aura(pet_cancel_aura_packet(creature_guid, spell_id))
-            .await;
-
-        assert!(!canonical_creature_has_applied_aura(
-            &canonical,
-            creature_guid,
-            aura
-        ));
-        assert!(send_rx.is_empty());
-    }
-
-    fn install_canonical_totem_for_session(
-        session: &mut crate::session::WorldSession,
-        canonical: &SharedCanonicalMapManager,
-        slot: usize,
-        is_totem: bool,
-    ) -> (ObjectGuid, ObjectGuid) {
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let totem_guid =
-            ObjectGuid::create_world_object(HighGuid::Creature, 0, 1, 571, 0, 777, slot as i64);
-        install_canonical_player(session, canonical, player_guid);
-        set_canonical_player_summon_slot(canonical, player_guid, slot, totem_guid);
-        add_canonical_test_creature_on_map(
-            canonical,
-            totem_guid,
-            Position::new(10.5, 20.5, 30.0, 0.0),
-            571,
-            0,
-            is_totem,
-        );
-        (player_guid, totem_guid)
-    }
-
-    #[tokio::test]
-    async fn totem_destroyed_despawns_matching_totem_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let slot = wow_entities::UNIT_SUMMON_SLOT_TOTEM + 2;
-        let (player_guid, totem_guid) =
-            install_canonical_totem_for_session(&mut session, &canonical, slot, true);
-
-        session
-            .handle_totem_destroyed(totem_destroyed_packet(2, totem_guid))
-            .await;
-
-        assert!(!canonical_creature_exists(&canonical, totem_guid));
-        assert_eq!(
-            canonical_player_summon_slot(&canonical, player_guid, slot),
-            ObjectGuid::EMPTY
-        );
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn totem_destroyed_empty_guid_matches_slot_totem_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let slot = wow_entities::UNIT_SUMMON_SLOT_TOTEM + 1;
-        let (player_guid, totem_guid) =
-            install_canonical_totem_for_session(&mut session, &canonical, slot, true);
-
-        session
-            .handle_totem_destroyed(totem_destroyed_packet(1, ObjectGuid::EMPTY))
-            .await;
-
-        assert!(!canonical_creature_exists(&canonical, totem_guid));
-        assert_eq!(
-            canonical_player_summon_slot(&canonical, player_guid, slot),
-            ObjectGuid::EMPTY
-        );
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn totem_destroyed_mismatched_guid_preserves_totem_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let slot = wow_entities::UNIT_SUMMON_SLOT_TOTEM;
-        let (player_guid, totem_guid) =
-            install_canonical_totem_for_session(&mut session, &canonical, slot, true);
-        let other_guid =
-            ObjectGuid::create_world_object(HighGuid::Creature, 0, 1, 571, 0, 777, 901);
-
-        session
-            .handle_totem_destroyed(totem_destroyed_packet(0, other_guid))
-            .await;
-
-        assert!(canonical_creature_exists(&canonical, totem_guid));
-        assert_eq!(
-            canonical_player_summon_slot(&canonical, player_guid, slot),
-            totem_guid
-        );
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn totem_destroyed_remote_control_preserves_totem_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let slot = wow_entities::UNIT_SUMMON_SLOT_TOTEM;
-        let (player_guid, totem_guid) =
-            install_canonical_totem_for_session(&mut session, &canonical, slot, true);
-        let controlled_guid =
-            ObjectGuid::create_world_object(HighGuid::Creature, 0, 1, 571, 0, 777, 902);
-        session.set_player_moved_unit_guid_like_cpp(controlled_guid);
-
-        session
-            .handle_totem_destroyed(totem_destroyed_packet(0, totem_guid))
-            .await;
-
-        assert!(canonical_creature_exists(&canonical, totem_guid));
-        assert_eq!(
-            canonical_player_summon_slot(&canonical, player_guid, slot),
-            totem_guid
-        );
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn totem_destroyed_out_of_range_slot_preserves_totem_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let slot = wow_entities::UNIT_SUMMON_SLOT_TOTEM;
-        let (player_guid, totem_guid) =
-            install_canonical_totem_for_session(&mut session, &canonical, slot, true);
-
-        session
-            .handle_totem_destroyed(totem_destroyed_packet(4, totem_guid))
-            .await;
-
-        assert!(canonical_creature_exists(&canonical, totem_guid));
-        assert_eq!(
-            canonical_player_summon_slot(&canonical, player_guid, slot),
-            totem_guid
-        );
-        assert!(send_rx.is_empty());
-    }
-
-    #[tokio::test]
-    async fn totem_destroyed_non_totem_creature_preserves_slot_like_cpp() {
-        let (mut session, send_rx) = make_session();
-        let canonical = shared_canonical_map_manager();
-        let slot = wow_entities::UNIT_SUMMON_SLOT_TOTEM;
-        let (player_guid, totem_guid) =
-            install_canonical_totem_for_session(&mut session, &canonical, slot, false);
-
-        session
-            .handle_totem_destroyed(totem_destroyed_packet(0, totem_guid))
-            .await;
-
-        assert!(canonical_creature_exists(&canonical, totem_guid));
-        assert_eq!(
-            canonical_player_summon_slot(&canonical, player_guid, slot),
-            totem_guid
-        );
-        assert!(send_rx.is_empty());
-    }
-
-    #[test]
-    fn open_item_wrapped_without_has_loot_uses_gift_row_like_cpp() {
-        let item_guid = ObjectGuid::create_item(1, 900);
-        let owner_guid = ObjectGuid::create_player(1, 42);
-        let gift_creator = ObjectGuid::create_player(1, 77);
-        let mut item = Item::new(0);
-        item.initialize_created_state(ItemCreateInfo {
-            guid: item_guid,
-            item_id: 100,
-            context: ItemContext::None,
-            owner: Some(owner_guid),
-            max_durability: 30,
-            expiration: 0,
-            spell_charges: [0; MAX_ITEM_SPELLS],
-        });
-        item.force_state(ItemUpdateState::Unchanged);
-        item.set_gift_creator(gift_creator);
-        item.replace_all_item_flags(ItemFieldFlags::WRAPPED);
-        item.set_durability(25);
-
-        let durability = apply_wrapped_gift_transform_like_cpp(
-            &mut item,
-            200,
-            ItemFieldFlags::SOULBOUND.bits(),
-            20,
-        );
-
-        assert_eq!(durability, 25);
-        assert_eq!(item.object().entry(), 200);
-        assert_eq!(item.data().gift_creator, ObjectGuid::EMPTY);
-        assert_eq!(item.item_flags_bits(), ItemFieldFlags::SOULBOUND.bits());
-        assert_eq!(item.data().max_durability, 20);
-        assert_eq!(item.data().durability, 25);
-        assert_eq!(item.update_state(), ItemUpdateState::Changed);
-        assert!(!item.is_wrapped());
-    }
-
-    #[test]
-    fn open_item_wrapped_gift_with_zero_durability_stays_zero_like_cpp() {
-        let item_guid = ObjectGuid::create_item(1, 901);
-        let owner_guid = ObjectGuid::create_player(1, 42);
-        let mut item = Item::new(0);
-        item.initialize_created_state(ItemCreateInfo {
-            guid: item_guid,
-            item_id: 100,
-            context: ItemContext::None,
-            owner: Some(owner_guid),
-            max_durability: 30,
-            expiration: 0,
-            spell_charges: [0; MAX_ITEM_SPELLS],
-        });
-        item.force_state(ItemUpdateState::Unchanged);
-        item.replace_all_item_flags(ItemFieldFlags::WRAPPED);
-        item.set_durability(0);
-
-        let durability = apply_wrapped_gift_transform_like_cpp(
-            &mut item,
-            200,
-            ItemFieldFlags::SOULBOUND.bits(),
-            20,
-        );
-
-        assert_eq!(durability, 0);
-        assert_eq!(item.data().max_durability, 20);
-        assert_eq!(item.data().durability, 0);
-        assert_eq!(item.update_state(), ItemUpdateState::Changed);
-        assert!(!item.is_wrapped());
-    }
-
-    #[test]
-    fn item_money_loot_generation_matches_cpp_boundary_branches() {
-        let mut rng = StdRng::seed_from_u64(0xC0FFEE);
-
-        assert_eq!(
-            wow_loot::generate_money_loot_with_rate_like_cpp(0, 0, 1.0, &mut rng),
-            0
-        );
-        let swapped = normalize_item_money_loot_bounds_like_cpp(120, 100);
-        assert_eq!(swapped, (100, 120));
-        let swapped_roll =
-            wow_loot::generate_money_loot_with_rate_like_cpp(swapped.0, swapped.1, 1.0, &mut rng);
-        assert!((100..=120).contains(&swapped_roll));
-        assert_eq!(
-            wow_loot::generate_money_loot_with_rate_like_cpp(100, 100, 1.0, &mut rng),
-            100
-        );
-        let swapped_rate_roll =
-            wow_loot::generate_money_loot_with_rate_like_cpp(swapped.0, swapped.1, 2.5, &mut rng);
-        assert!((250..=300).contains(&swapped_rate_roll));
-
-        let small_range = wow_loot::generate_money_loot_with_rate_like_cpp(100, 200, 1.0, &mut rng);
-        assert!((100..=200).contains(&small_range));
-
-        let wide_range =
-            wow_loot::generate_money_loot_with_rate_like_cpp(1_000, 100_000, 1.0, &mut rng);
-        assert_eq!(wide_range & 0xFF, 0);
-        assert!((((1_000 >> 8) << 8)..=((100_000 >> 8) << 8)).contains(&wide_range));
-    }
-
-    #[tokio::test]
-    async fn item_template_addon_port_preserves_found_rows_and_request_order() {
-        let fixture = ItemTemplateAddonCatalogPortFixtureLikeCpp::new(
-            [ItemTemplateAddonMoneyOutcomeLikeCpp::Found(
-                wow_persistence::ItemTemplateAddonMoneyRowLikeCpp {
-                    min_money: Some(120),
-                    max_money: Some(100),
-                },
-            )],
-            [ItemTemplateAddonLootMetadataOutcomeLikeCpp::Found(
-                wow_persistence::ItemTemplateAddonLootMetadataRowLikeCpp {
-                    flags_cu: ITEM_FLAGS_CU_FOLLOW_LOOT_RULES_LIKE_CPP,
-                    quest_log_item_id: 77,
-                },
-            )],
-        );
-        let (mut session, _send_rx) = make_session();
-        session.set_item_template_addon_catalog_persistence_port_like_cpp(fixture.clone());
-
-        assert_eq!(
-            session
-                .load_item_template_addon_money_loot_like_cpp(1001)
-                .await,
-            (100, 120)
-        );
-        assert_eq!(
-            session
-                .load_item_template_addon_loot_metadata_like_cpp(1002)
-                .await,
-            ItemTemplateAddonLootMetadataLikeCpp {
-                flags_cu: ITEM_FLAGS_CU_FOLLOW_LOOT_RULES_LIKE_CPP,
-                quest_log_item_id: 77,
-            }
-        );
-        assert_eq!(
-            *fixture.money_requests.lock().unwrap(),
-            [ItemTemplateAddonCatalogRequestLikeCpp { item_entry: 1001 }]
-        );
-        assert_eq!(
-            *fixture.metadata_requests.lock().unwrap(),
-            [ItemTemplateAddonCatalogRequestLikeCpp { item_entry: 1002 }]
-        );
-    }
-
-    #[tokio::test]
-    async fn item_template_addon_port_preserves_absent_malformed_and_failed_defaults() {
-        let fixture = ItemTemplateAddonCatalogPortFixtureLikeCpp::new(
-            [
-                ItemTemplateAddonMoneyOutcomeLikeCpp::Found(
-                    wow_persistence::ItemTemplateAddonMoneyRowLikeCpp {
-                        min_money: Some(10),
-                        max_money: None,
-                    },
-                ),
-                ItemTemplateAddonMoneyOutcomeLikeCpp::Missing,
-                ItemTemplateAddonMoneyOutcomeLikeCpp::Failed {
-                    reason: "world read failed".into(),
-                },
-            ],
-            [
-                ItemTemplateAddonLootMetadataOutcomeLikeCpp::Missing,
-                ItemTemplateAddonLootMetadataOutcomeLikeCpp::Failed {
-                    reason: "world read failed".into(),
-                },
-            ],
-        );
-        let (mut session, _send_rx) = make_session();
-        session.set_item_template_addon_catalog_persistence_port_like_cpp(fixture);
-
-        for item_entry in [2001, 2002, 2003] {
-            assert_eq!(
-                session
-                    .load_item_template_addon_money_loot_like_cpp(item_entry)
-                    .await,
-                (0, 0)
-            );
-        }
-        for item_entry in [2004, 2005] {
-            assert_eq!(
-                session
-                    .load_item_template_addon_loot_metadata_like_cpp(item_entry)
-                    .await,
-                ItemTemplateAddonLootMetadataLikeCpp::default()
-            );
-        }
-
-        let (session_without_port, _send_rx) = make_session();
-        assert_eq!(
-            session_without_port
-                .load_item_template_addon_money_loot_like_cpp(2006)
-                .await,
-            (0, 0)
-        );
-        assert_eq!(
-            session_without_port
-                .load_item_template_addon_loot_metadata_like_cpp(2007)
-                .await,
-            ItemTemplateAddonLootMetadataLikeCpp::default()
-        );
-    }
-
-    #[test]
-    fn loot_rate_helpers_match_cpp_roll_inputs() {
-        let mut rng = StdRng::seed_from_u64(0xA11CE);
-        assert!(roll_chance_with_rate_like_cpp(100.0, 0.0, &mut rng));
-        assert!(roll_chance_with_rate_like_cpp(60.0, 2.0, &mut rng));
-        assert_eq!(referenced_loot_max_count_like_cpp(3, 2.0), 6);
-        assert_eq!(referenced_loot_max_count_like_cpp(3, 0.5), 1);
-    }
-
-    #[test]
-    fn plain_item_loot_template_validation_matches_cpp_basic_guards() {
-        assert!(loot_template_plain_row_can_roll_like_cpp(
-            25, 100.0, false, 1, 1, 3, true, true
-        ));
-        assert!(!loot_template_plain_row_can_roll_like_cpp(
-            0, 100.0, false, 1, 1, 3, true, true
-        ));
-        assert!(!loot_template_plain_row_can_roll_like_cpp(
-            25, 0.0, false, 1, 1, 3, true, true
-        ));
-        assert!(loot_template_plain_row_can_roll_like_cpp(
-            25, 100.0, true, 1, 1, 3, true, true
-        ));
-        assert!(!loot_template_plain_row_can_roll_like_cpp(
-            25, 100.0, true, 1, 1, 3, true, false
-        ));
-        assert!(!loot_template_plain_row_can_roll_like_cpp(
-            25, 100.0, false, 0, 1, 3, true, true
-        ));
-        assert!(!loot_template_plain_row_can_roll_like_cpp(
-            25, 100.0, false, 1, 0, 3, true, true
-        ));
-        assert!(!loot_template_plain_row_can_roll_like_cpp(
-            25, 100.0, false, 1, 4, 3, true, true
-        ));
-        assert!(!loot_template_plain_row_can_roll_like_cpp(
-            25, 100.0, false, 1, 1, 3, false, true
-        ));
-    }
-
-    #[test]
-    fn item_loot_quest_status_custom_metadata_matches_cpp_gate() {
-        assert!(item_loot_quest_status_allows_like_cpp(
-            true, true, true, false
-        ));
-        assert!(item_loot_quest_status_allows_like_cpp(
-            false, false, false, false
-        ));
-        assert!(!item_loot_quest_status_allows_like_cpp(
-            false, true, false, false
-        ));
-        assert!(!item_loot_quest_status_allows_like_cpp(
-            false, false, true, false
-        ));
-        assert!(item_loot_quest_status_allows_like_cpp(
-            false, true, false, true
-        ));
-
-        let metadata = ItemTemplateAddonLootMetadataLikeCpp {
-            flags_cu: ITEM_FLAGS_CU_IGNORE_QUEST_STATUS_LIKE_CPP,
-            quest_log_item_id: 25,
-        };
-        assert!(metadata.ignores_quest_status());
-        assert_eq!(metadata.quest_log_item_id, 25);
-    }
-
-    #[test]
-    fn loot_entry_follow_rules_metadata_matches_cpp_constructor() {
-        let normal = loot_entry_flags_for_row_metadata_like_cpp(
-            false,
-            ItemFlags::empty(),
-            ItemTemplateAddonLootMetadataLikeCpp::default(),
-        );
-        assert!(normal.follow_loot_rules);
-        assert!(!normal.needs_quest);
-        assert!(!normal.freeforall);
-
-        let quest_only_without_custom = loot_entry_flags_for_row_metadata_like_cpp(
-            true,
-            ItemFlags::empty(),
-            ItemTemplateAddonLootMetadataLikeCpp::default(),
-        );
-        assert!(!quest_only_without_custom.follow_loot_rules);
-        assert!(quest_only_without_custom.needs_quest);
-
-        let quest_only_following_rules = loot_entry_flags_for_row_metadata_like_cpp(
-            true,
-            ItemFlags::MULTI_DROP,
-            ItemTemplateAddonLootMetadataLikeCpp {
-                flags_cu: ITEM_FLAGS_CU_FOLLOW_LOOT_RULES_LIKE_CPP,
-                quest_log_item_id: 0,
-            },
-        );
-        assert!(quest_only_following_rules.follow_loot_rules);
-        assert!(quest_only_following_rules.freeforall);
-        assert!(quest_only_following_rules.needs_quest);
-    }
-
-    #[test]
-    fn loot_entry_free_for_all_view_metadata_matches_cpp() {
-        let player_guid = ObjectGuid::create_player(1, 42);
-        let other_guid = ObjectGuid::create_player(1, 43);
-        let normal = LootEntry {
-            loot_list_id: 0,
-            item_id: 25,
-            quantity: 1,
-            random_properties_id: 0,
-            random_properties_seed: 0,
-            item_context: 0,
-            flags: LootEntryFlags::default(),
-            allowed_looters: vec![player_guid],
-            roll_winner: ObjectGuid::EMPTY,
-            ffa_looted_by: Vec::new(),
-            taken: false,
-        };
-        assert_eq!(normal.free_for_all_ui_type_like_cpp(), 4);
-        assert!(normal.is_over_threshold_like_cpp());
-        assert!(normal.visible_in_represented_free_for_all_view_like_cpp(player_guid));
-        assert!(!normal.visible_in_represented_free_for_all_view_like_cpp(other_guid));
-
-        let ffa = LootEntry {
-            flags: LootEntryFlags {
-                freeforall: true,
-                needs_quest: true,
-                ..Default::default()
-            },
-            ..normal.clone()
-        };
-        assert_eq!(ffa.free_for_all_ui_type_like_cpp(), 4);
-        assert!(!ffa.is_over_threshold_like_cpp());
-        assert!(ffa.flags.needs_quest);
-        assert!(!ffa.is_looted_for_player_like_cpp(player_guid));
-        let mut looted_ffa = ffa.clone();
-        looted_ffa.add_allowed_looter_like_cpp(other_guid);
-        looted_ffa.mark_looted_for_player_like_cpp(player_guid);
-        assert!(!looted_ffa.taken);
-        assert!(looted_ffa.is_looted_for_player_like_cpp(player_guid));
-        assert!(!looted_ffa.is_looted_for_player_like_cpp(other_guid));
-        assert!(!looted_ffa.fully_looted_like_cpp());
-        looted_ffa.mark_looted_for_player_like_cpp(other_guid);
-        assert!(looted_ffa.fully_looted_like_cpp());
-
-        let under_threshold = LootEntry {
-            flags: LootEntryFlags {
-                under_threshold: true,
-                ..Default::default()
-            },
-            ..normal
-        };
-        assert!(!under_threshold.is_over_threshold_like_cpp());
-
-        let taken = LootEntry {
-            taken: true,
-            ..under_threshold
-        };
-        assert!(!taken.visible_in_represented_free_for_all_view_like_cpp(player_guid));
-    }
-
-    fn condition(
-        else_group: u32,
-        condition_type_or_reference: i32,
-        value1: u32,
-        negative: bool,
-    ) -> LootConditionRowLikeCpp {
-        LootConditionRowLikeCpp {
-            else_group,
-            condition_type_or_reference,
-            condition_target: 0,
-            value1,
-            value2: 0,
-            value3: 0,
-            string_value1: String::new(),
-            negative,
-            script_name: String::new(),
-        }
-    }
-
-    #[test]
-    fn loot_conditions_else_group_and_negative_match_cpp() {
-        let conditions = vec![
-            condition(0, 25, 100, false),
-            condition(0, 25, 200, false),
-            condition(1, 25, 300, true),
-        ];
-
-        assert!(loot_conditions_allow_player_like_cpp_representable(
-            &conditions,
-            |condition| Some(condition.value1 == 100 || condition.value1 == 200),
-        ));
-        assert!(loot_conditions_allow_player_like_cpp_representable(
-            &conditions,
-            |condition| Some(condition.value1 == 100),
-        ));
-        assert!(!loot_conditions_allow_player_like_cpp_representable(
-            &conditions,
-            |condition| Some(condition.value1 == 300),
-        ));
-
-        let mut scripted = condition(0, 25, 100, false);
-        scripted.script_name = "npc_custom".to_string();
-        assert!(!loot_conditions_allow_player_like_cpp_representable(
-            &[scripted],
-            |_| Some(true),
-        ));
-        assert!(loot_conditions_allow_player_like_cpp_representable(
-            &[condition(0, -1, 100, false)],
-            |_| Some(true),
-        ));
-    }
-
-    #[test]
-    fn loot_condition_compare_and_quest_state_match_cpp_values() {
-        assert_eq!(condition_compare_values_like_cpp(0, 10, 10), Some(true));
-        assert_eq!(condition_compare_values_like_cpp(1, 11, 10), Some(true));
-        assert_eq!(condition_compare_values_like_cpp(2, 9, 10), Some(true));
-        assert_eq!(condition_compare_values_like_cpp(3, 10, 10), Some(true));
-        assert_eq!(condition_compare_values_like_cpp(4, 10, 10), Some(true));
-        assert_eq!(condition_compare_values_like_cpp(5, 10, 10), None);
-
-        assert_eq!(player_quest_status_mask_like_cpp(None, false), 1);
-        assert_eq!(player_quest_status_mask_like_cpp(Some(2), false), 2);
-        assert_eq!(player_quest_status_mask_like_cpp(Some(1), false), 8);
-        assert_eq!(player_quest_status_mask_like_cpp(Some(3), false), 32);
-        assert_eq!(player_quest_status_mask_like_cpp(Some(1), true), 64);
-
-        assert_eq!(player_class_mask_like_cpp(13), Some(1 << 12));
-        assert_eq!(player_class_mask_like_cpp(14), None);
-        assert_eq!(player_race_mask_like_cpp(34), Some(1 << 11));
-        assert_eq!(player_race_mask_like_cpp(35), Some(1 << 12));
-        assert_eq!(player_race_mask_like_cpp(52), Some(1 << 16));
-        assert_eq!(player_race_mask_like_cpp(70), Some(1 << 15));
-        assert_eq!(player_race_mask_like_cpp(33), None);
-    }
-
-    #[test]
-    fn add_loot_item_stacks_caps_like_cpp_max_nr_loot_items() {
-        let mut loot_items = Vec::new();
-        add_loot_item_stacks_like_cpp(&mut loot_items, 25, 45, 20, Default::default());
-        assert_eq!(loot_items.len(), 3);
-        assert_eq!(loot_items[0].quantity, 20);
-        assert_eq!(loot_items[1].quantity, 20);
-        assert_eq!(loot_items[2].quantity, 5);
-        assert_eq!(loot_items[2].loot_list_id, 2);
-        assert_eq!(loot_items[2].random_properties_id, 0);
-        assert_eq!(loot_items[2].random_properties_seed, 0);
-        assert_eq!(loot_items[2].item_context, 0);
-
-        let mut capped = Vec::new();
-        add_loot_item_stacks_like_cpp(&mut capped, 25, 100, 1, Default::default());
-        assert_eq!(capped.len(), 18);
-        assert_eq!(capped[17].loot_list_id, 17);
-    }
-
-    #[test]
-    fn add_loot_template_row_item_uses_caller_rng_like_cpp_urand_count() {
-        let row = LootTemplateRow {
-            item_id: 25,
-            reference: 0,
-            chance: 100.0,
-            needs_quest: false,
-            loot_mode: LOOT_MODE_DEFAULT_LIKE_CPP,
-            group_id: 0,
-            min_count: 2,
-            max_count: 7,
-            conditions: Vec::new(),
-        };
-        let mut expected_rng = StdRng::seed_from_u64(0x5151);
-        let expected_count = expected_rng.gen_range(2..=7);
-
-        let mut rng = StdRng::seed_from_u64(0x5151);
-        let mut loot_items = Vec::new();
-        add_loot_template_row_item_like_cpp(
-            &mut loot_items,
-            &row,
-            Default::default(),
-            |_| 20,
-            &mut rng,
-        );
-
-        assert_eq!(loot_items.len(), 1);
-        assert_eq!(loot_items[0].quantity, expected_count);
-    }
-
-    #[test]
-    fn reference_loot_template_validation_matches_cpp_basic_guards() {
-        assert!(loot_template_reference_row_can_roll_like_cpp(
-            10, 100.0, 1, 1
-        ));
-        assert!(loot_template_reference_row_can_roll_like_cpp(
-            10, 0.0000001, 1, 1
-        ));
-        assert!(!loot_template_reference_row_can_roll_like_cpp(
-            0, 100.0, 1, 1
-        ));
-        assert!(!loot_template_reference_row_can_roll_like_cpp(
-            10, 0.0, 1, 1
-        ));
-        assert!(!loot_template_reference_row_can_roll_like_cpp(
-            10, 100.0, 0, 1
-        ));
-        assert!(!loot_template_reference_row_can_roll_like_cpp(
-            10, 100.0, 1, 0
-        ));
-    }
-
-    #[test]
-    fn grouped_loot_template_roll_matches_cpp_explicit_then_equal_order() {
-        assert!(loot_template_group_row_can_roll_like_cpp(
-            25, 0.0, false, 1, 1, 1, true, true
-        ));
-        assert!(loot_template_group_row_can_roll_like_cpp(
-            25, 0.0, true, 1, 1, 1, true, true
-        ));
-        assert!(!loot_template_group_row_can_roll_like_cpp(
-            25, 0.0, true, 1, 1, 1, true, false
-        ));
-        assert!(!loot_template_group_row_can_roll_like_cpp(
-            25, 0.0000001, false, 1, 1, 1, true, true
-        ));
-
-        let rows = vec![
-            LootTemplateRow {
-                item_id: 25,
-                reference: 0,
-                chance: 100.0,
-                needs_quest: false,
-                loot_mode: 1,
-                group_id: 1,
-                min_count: 1,
-                max_count: 1,
-                conditions: Vec::new(),
-            },
-            LootTemplateRow {
-                item_id: 26,
-                reference: 0,
-                chance: 0.0,
-                needs_quest: false,
-                loot_mode: 1,
-                group_id: 1,
-                min_count: 1,
-                max_count: 1,
-                conditions: Vec::new(),
-            },
-        ];
-        let mut rng = StdRng::seed_from_u64(0xBEEF);
-        let selected =
-            roll_group_loot_row_like_cpp(&rows, 1, |_| true, |_| true, |_| 1.0, &mut rng).unwrap();
-        assert_eq!(selected.item_id, 25);
-
-        let equal_rows = vec![
-            LootTemplateRow {
-                chance: 0.0,
-                item_id: 25,
-                ..rows[0].clone()
-            },
-            LootTemplateRow {
-                chance: 0.0,
-                item_id: 26,
-                ..rows[1].clone()
-            },
-        ];
-        let mut rng = StdRng::seed_from_u64(0xBEEF);
-        let selected =
-            roll_group_loot_row_like_cpp(&equal_rows, 1, |_| true, |_| true, |_| 1.0, &mut rng)
-                .unwrap();
-        assert!([25, 26].contains(&selected.item_id));
-    }
-
-    #[test]
-    fn open_item_stored_loot_preserves_random_properties_and_context_like_cpp() {
-        assert!(stored_item_row_can_load_like_cpp_representable(
-            25, 2, 7, false, false, -77, 456, 2, true
-        ));
-        assert!(stored_item_row_can_load_like_cpp_representable(
-            25, 2, 7, false, true, -77, 456, 2, true
-        ));
-
-        let entry = LootEntry {
-            loot_list_id: 7,
-            item_id: 25,
-            quantity: 2,
-            random_properties_id: -77,
-            random_properties_seed: 456,
-            item_context: 2,
-            flags: LootEntryFlags::default(),
-            allowed_looters: Vec::new(),
-            roll_winner: ObjectGuid::EMPTY,
-            ffa_looted_by: Vec::new(),
-            taken: false,
-        };
-        let response_item = wow_packet::packets::loot::LootItemData {
-            item_type: 0,
-            ui_type: entry.free_for_all_ui_type_like_cpp(),
-            can_trade_to_tap_list: false,
-            loot: wow_packet::packets::item::ItemInstance {
-                item_id: entry.item_id as i32,
-                ..wow_packet::packets::item::ItemInstance::default()
-            },
-            loot_list_id: entry.loot_list_id,
-            quantity: entry.quantity,
-            loot_item_type: 0,
-        };
-        assert_eq!(entry.random_properties_id, -77);
-        assert_eq!(entry.random_properties_seed, 456);
-        assert_eq!(entry.item_context, 2);
-        assert_eq!(response_item.loot.item_id, 25);
-        assert!(response_item.loot.item_bonus.is_none());
-
-        assert!(!stored_item_row_can_load_like_cpp_representable(
-            25, 2, 256, false, false, 0, 0, 0, true
-        ));
-        assert!(!stored_item_row_can_load_like_cpp_representable(
-            25, 2, 7, true, false, 0, 0, 0, true
-        ));
-    }
-
-    #[test]
-    fn stored_loot_item_persistence_skips_missing_template_and_currency_tokens_like_cpp() {
-        // template missing -> no persist
-        assert!(!stored_loot_item_should_persist_like_cpp(
-            false,
-            BagFamilyMask::NONE
-        ));
-
-        // normal template -> persist
-        assert!(stored_loot_item_should_persist_like_cpp(
-            true,
-            BagFamilyMask::NONE
-        ));
-
-        // currency token -> no persist (C++ ItemTemplate::IsCurrencyToken)
-        assert!(!stored_loot_item_should_persist_like_cpp(
-            true,
-            BagFamilyMask::CURRENCY_TOKENS
-        ));
-
-        // currency token combined with other families still no persist
-        assert!(!stored_loot_item_should_persist_like_cpp(
-            true,
-            BagFamilyMask::CURRENCY_TOKENS | BagFamilyMask::HERBS
-        ));
-    }
-}
+#[path = "spell/tests/mod.rs"]
+mod tests;
